Refactor: Move around a lot of files. Unify build and switch to using dts-bundle-generator
