+++ src/engine/GoMath.ts
