+++ src/engine/GobanError.ts
