+++ src/engine/README.md
@@ -1,0 +1,5 @@
+The goban engine module contains all of the logic for playing the game and
+communicating with the Online-Go.com game servers. 
+
+The code in this module **MUST** be able to be compiled and largely operate
+in both browser and node environments.
