+++ src/engine/StoneString.ts
