+++ src/engine/StoneStringBuilder.ts
