+++ src/engine/ownership_estimators/wasm_estimator.ts
