+++ src/engine/util.ts
