+++ src/goban/GobanCanvas.ts
@@ -14,37 +14,38 @@
  * limitations under the License.
  */
 
-import { JGOF, JGOFIntersection, JGOFNumericPlayerColor } from "./JGOF";
+import { JGOF, JGOFIntersection, JGOFNumericPlayerColor } from "engine/JGOF";
 
-import { AdHocFormat } from "./AdHocFormat";
+import { AdHocFormat } from "engine/AdHocFormat";
 
-//import { GobanCore, GobanSelectedThemes, GobanMetrics, GOBAN_FONT } from "./GobanCore";
-import { GobanCore, GobanConfig, GobanSelectedThemes, GobanMetrics } from "./GobanCore";
-import { GoEngine } from "./GoEngine";
-import * as GoMath from "./GoMath";
-import { MoveTree } from "./MoveTree";
+import { GobanCore, GobanConfig, GobanSelectedThemes, GobanMetrics, GOBAN_FONT } from "./GobanCore";
+import { GoEngine } from "engine/GoEngine";
+import * as GoMath from "engine/GoMath";
+import { MoveTree } from "engine/MoveTree";
 import { GoTheme } from "./GoTheme";
 import { GoThemes } from "./GoThemes";
-import { MoveTreePenMarks } from "./MoveTree";
-import { getRelativeEventPosition } from "./canvas_utils";
-import { _ } from "./translate";
-import { formatMessage, MessageID } from "./messages";
-import { color_blend, getRandomInt } from "./util";
-import { callbacks } from "./callbacks";
-
-//import { GobanCanvasConfig, GobanCanvasInterface } from "./GobanCanvas";
+import { MoveTreePenMarks } from "engine/MoveTree";
+import {
+    createDeviceScaledCanvas,
+    resizeDeviceScaledCanvas,
+    allocateCanvasOrError,
+    getRelativeEventPosition,
+} from "./canvas_utils";
+import { _ } from "engine/translate";
+import { formatMessage, MessageID } from "engine/messages";
+import { color_blend, getRandomInt } from "engine/util";
+import { callbacks } from "../engine/callbacks";
 
 const __theme_cache: {
-    [color: string]: { [name: string]: { [size: string]: any } };
+    [bw: string]: { [name: string]: { [size: string]: any } };
 } = {
     black: {},
     white: {},
 };
-//const __theme_defs_cache: { [radius: number]: SVGDefsElement } = {};
 
 declare let ResizeObserver: any;
 
-export interface GobanSVGConfig extends GobanConfig {
+export interface GobanCanvasConfig extends GobanConfig {
     board_div?: HTMLElement;
     title_div?: HTMLElement;
     move_tree_container?: HTMLElement;
@@ -62,8 +63,7 @@ interface ViewPortInterface {
 
 const HOT_PINK = "#ff69b4";
 
-//interface GobanCanvasInterface {
-interface GobanSVGInterface {
+export interface GobanCanvasInterface {
     engine: GoEngine;
     move_tree_container?: HTMLElement;
 
@@ -74,7 +74,7 @@ interface GobanSVGInterface {
     setByoYomiLabel(label: string): void;
     setLastMoveOpacity(opacity: number): void;
 
-    move_tree_bindEvents(svg: SVGElement): void;
+    move_tree_bindCanvasEvents(canvas: HTMLCanvasElement): void;
     move_tree_redraw(no_warp?: boolean): void;
     setMoveTreeContainer(container: HTMLElement): void;
 
@@ -90,13 +90,10 @@ interface GobanSVGInterface {
     destroy(): void;
 }
 
-export class GobanSVG extends GobanCore implements GobanSVGInterface {
+export class GobanCanvas extends GobanCore implements GobanCanvasInterface {
     public engine: GoEngine;
     //private board_div: HTMLElement;
-    private svg: SVGElement;
-    private svg_defs: SVGDefsElement;
-    //private loaded_pre_rendered_stones: { [radius: number]: boolean } = {};
-    public event_layer: HTMLDivElement;
+    private board: HTMLCanvasElement;
     private __set_board_height: number = -1;
     private __set_board_width: number = -1;
     private ready_to_draw: boolean = false;
@@ -104,23 +101,19 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
     private message_td?: HTMLElement;
     private message_text?: HTMLDivElement;
     private message_timeout?: number;
+    private shadow_layer?: HTMLCanvasElement;
+    private shadow_ctx?: CanvasRenderingContext2D;
     private handleShiftKey: (ev: KeyboardEvent) => void;
-
-    private lines_layer?: SVGGraphicsElement;
-    private coordinate_labels_layer?: SVGGraphicsElement;
-    private grid: Array<Array<SVGGraphicsElement>> = [];
-    private grid_layer?: SVGGraphicsElement;
-    private shadow_grid: Array<Array<SVGElement | undefined>> = [];
-    private shadow_layer?: SVGGraphicsElement;
-    private pen_layer?: SVGGraphicsElement;
 
     private last_move_opacity: number = 1;
     public move_tree_container?: HTMLElement;
     private move_tree_inner_container?: HTMLDivElement;
-    private move_tree_svg?: SVGElement;
+    private move_tree_canvas?: HTMLCanvasElement;
 
+    private __borders_initialized: boolean = false;
     private autoplaying_puzzle_move: boolean = false;
     private byoyomi_label: string = "";
+    private ctx: CanvasRenderingContext2D;
     private current_pen_mark?: { color: string; points: [number, number] };
     private labeling_mode?: "put" | "clear";
     private last_label_position: { i: number; j: number } = { i: NaN, j: NaN };
@@ -131,6 +124,8 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
     private analysis_scoring_last_position: { i: number; j: number } = { i: NaN, j: NaN };
 
     private drawing_enabled: boolean = true;
+    private pen_ctx?: CanvasRenderingContext2D;
+    private pen_layer?: HTMLCanvasElement;
     protected title_div?: HTMLElement;
 
     private themes: GobanSelectedThemes = {
@@ -153,33 +148,35 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
     private theme_white_stones: Array<any> = [];
     private theme_white_text_color: string = HOT_PINK;
 
-    constructor(config: GobanSVGConfig, preloaded_data?: AdHocFormat | JGOF) {
+    constructor(config: GobanCanvasConfig, preloaded_data?: AdHocFormat | JGOF) {
         /* TODO: Need to reconcile the clock fields before we can get rid of this `any` cast */
         super(config, preloaded_data as any);
 
+        // console.log("Goban canvas v 0.5.74.debug 5"); // GaJ: I use this to be sure I have linked & loaded the updates
         if (config.board_div) {
             this.parent = config["board_div"];
         } else {
             this.no_display = true;
-            // unattached div dangle prevent null pointer refs
-            this.parent = document.createElement("div");
+            this.parent =
+                document.createElement(
+                    "div",
+                ); /* let a div dangle in no-mans land to prevent null pointer refs */
         }
 
         this.title_div = config["title_div"];
-        this.svg = document.createElementNS("http://www.w3.org/2000/svg", "svg");
-        this.svg_defs = document.createElementNS("http://www.w3.org/2000/svg", "defs");
-        this.svg.appendChild(this.svg_defs);
+        this.board = createDeviceScaledCanvas(10, 10);
+        this.board.setAttribute("id", "board-canvas");
+        this.board.className = "StoneLayer";
         this.last_move_opacity = config["last_move_opacity"] ?? 1;
+        const ctx = this.board.getContext("2d", { willReadFrequently: true });
+        if (ctx) {
+            this.ctx = ctx;
+        } else {
+            throw new Error(`Failed to obtain drawing context for board`);
+        }
 
-        this.parent.appendChild(this.svg);
-        this.event_layer = document.createElement("div");
-        this.event_layer.style.position = "absolute";
-        this.event_layer.style.top = "0";
-        this.event_layer.style.right = "0";
-        this.event_layer.style.left = "0";
-        this.event_layer.style.bottom = "0";
-        this.parent.appendChild(this.event_layer);
-        this.bindPointerBindings(this.event_layer);
+        this.parent.appendChild(this.board);
+        this.bindPointerBindings(this.board);
 
         this.move_tree_container = config.move_tree_container;
 
@@ -231,24 +228,23 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
     }
 
     public enablePen(): void {
-        this.attachPenLayer();
+        this.attachPenCanvas();
     }
     public disablePen(): void {
-        this.detachPenLayer();
+        this.detachPenCanvas();
     }
 
     public destroy(): void {
         super.destroy();
 
-        this.clearMessage();
-
-        if (this.svg && this.svg.parentNode) {
-            this.svg.remove();
+        if (this.board && this.board.parentNode) {
+            this.board.parentNode.removeChild(this.board);
         }
         delete (this as any).board;
-        delete (this as any).svg;
+        delete (this as any).ctx;
 
-        this.detachPenLayer();
+        this.detachPenCanvas();
+        this.detachShadowLayer();
 
         if (this.message_timeout) {
             clearTimeout(this.message_timeout);
@@ -268,35 +264,82 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         delete this.message_text;
         delete this.move_tree_container;
         delete this.move_tree_inner_container;
-        delete this.move_tree_svg;
+        delete this.move_tree_canvas;
         delete this.title_div;
     }
-    private detachPenLayer(): void {
+    private detachShadowLayer(): void {
+        if (this.shadow_layer) {
+            if (this.shadow_layer.parentNode) {
+                this.shadow_layer.parentNode.removeChild(this.shadow_layer);
+            }
+            delete this.shadow_layer;
+            delete this.shadow_ctx;
+        }
+    }
+    private attachShadowLayer(): void {
+        if (!this.shadow_layer && this.parent) {
+            this.shadow_layer = createDeviceScaledCanvas(this.metrics.width, this.metrics.height);
+            this.shadow_layer.setAttribute("id", "shadow-canvas");
+            this.shadow_layer.className = "ShadowLayer";
+
+            try {
+                this.parent.insertBefore(this.shadow_layer, this.board);
+            } catch (e) {
+                // I'm not really sure how we ever get into this state, but sentry.io reports that we do
+                console.warn("Error inserting shadow layer before board");
+                console.warn(e);
+                try {
+                    this.parent.appendChild(this.shadow_layer);
+                } catch (e) {
+                    console.error(e);
+                }
+            }
+
+            const ctx = this.shadow_layer.getContext("2d", { willReadFrequently: true });
+            if (ctx) {
+                this.shadow_ctx = ctx;
+            } else {
+                //throw new Error(`Failed to obtain shadow layer drawing context`);
+                console.error(new Error(`Failed to obtain shadow layer drawing context`));
+                return;
+            }
+            this.bindPointerBindings(this.shadow_layer);
+        }
+    }
+    private detachPenCanvas(): void {
         if (this.pen_layer) {
-            this.pen_layer.remove();
+            if (this.pen_layer.parentNode) {
+                this.pen_layer.parentNode.removeChild(this.pen_layer);
+            }
             delete this.pen_layer;
+            delete this.pen_ctx;
         }
     }
-    private attachPenLayer(): void {
+    private attachPenCanvas(): void {
         if (!this.pen_layer) {
-            this.pen_layer = document.createElementNS("http://www.w3.org/2000/svg", "g");
-            this.pen_layer.setAttribute("id", "pen-svg");
-            this.pen_layer.setAttribute("class", "PenLayer");
-            this.pen_layer.setAttribute("width", this.metrics.width.toString());
-            this.pen_layer.setAttribute("height", this.metrics.height.toString());
-            this.svg.appendChild(this.pen_layer);
+            this.pen_layer = createDeviceScaledCanvas(this.metrics.width, this.metrics.height);
+            this.pen_layer.setAttribute("id", "pen-canvas");
+            this.pen_layer.className = "PenLayer";
+            this.parent.appendChild(this.pen_layer);
+            const ctx = this.pen_layer.getContext("2d", { willReadFrequently: true });
+            if (ctx) {
+                this.pen_ctx = ctx;
+            } else {
+                throw new Error(`Failed to obtain pen drawing context`);
+            }
+            this.bindPointerBindings(this.pen_layer);
         }
     }
-    private bindPointerBindings(div: HTMLDivElement): void {
+    private bindPointerBindings(canvas: HTMLCanvasElement): void {
         if (!this.interactive) {
             return;
         }
 
-        if (div.getAttribute("data-pointers-bound") === "true") {
+        if (canvas.getAttribute("data-pointers-bound") === "true") {
             return;
         }
 
-        div.setAttribute("data-pointers-bound", "true");
+        canvas.setAttribute("data-pointers-bound", "true");
 
         let dragging = false;
 
@@ -307,6 +350,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             const press_duration_ms = performance.now() - pointer_down_timestamp;
             try {
                 if (!dragging) {
+                    /* if we didn't start the click in the canvas, don't respond to it */
                     return;
                 }
                 let right_click = false;
@@ -320,7 +364,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 dragging = false;
 
                 if (this.scoring_mode) {
-                    const pos = getRelativeEventPosition(ev, this.parent);
+                    const pos = getRelativeEventPosition(ev);
                     const pt = this.xy2ij(pos.x, pos.y);
                     if (pt.i >= 0 && pt.i < this.width && pt.j >= 0 && pt.j < this.height) {
                         if (this.score_estimator) {
@@ -338,7 +382,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
                 if (ev.ctrlKey || ev.metaKey || ev.altKey) {
                     try {
-                        const pos = getRelativeEventPosition(ev, this.parent);
+                        const pos = getRelativeEventPosition(ev);
                         const pt = this.xy2ij(pos.x, pos.y);
                         if (callbacks.addCoordinatesToChatInput) {
                             callbacks.addCoordinatesToChatInput(
@@ -358,7 +402,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 } else if (this.mode === "analyze" && this.analyze_tool === "removal") {
                     this.onAnalysisToggleStoneRemoval(ev);
                 } else {
-                    const pos = getRelativeEventPosition(ev, this.parent);
+                    const pos = getRelativeEventPosition(ev);
                     const pt = this.xy2ij(pos.x, pos.y);
                     if (!double_clicked) {
                         last_click_square = pt;
@@ -441,7 +485,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
         let mouse_disabled: any = 0;
 
-        div.addEventListener("click", (ev) => {
+        canvas.addEventListener("click", (ev) => {
             if (!mouse_disabled) {
                 dragging = true;
                 pointerUp(ev, false);
@@ -449,7 +493,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             ev.preventDefault();
             return false;
         });
-        div.addEventListener("dblclick", (ev) => {
+        canvas.addEventListener("dblclick", (ev) => {
             if (!mouse_disabled) {
                 dragging = true;
                 pointerUp(ev, true);
@@ -457,21 +501,21 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             ev.preventDefault();
             return false;
         });
-        div.addEventListener("mousedown", (ev) => {
+        canvas.addEventListener("mousedown", (ev) => {
             if (!mouse_disabled) {
                 pointerDown(ev);
             }
             ev.preventDefault();
             return false;
         });
-        div.addEventListener("mousemove", (ev) => {
+        canvas.addEventListener("mousemove", (ev) => {
             if (!mouse_disabled) {
                 pointerMove(ev);
             }
             ev.preventDefault();
             return false;
         });
-        div.addEventListener("mouseout", (ev) => {
+        canvas.addEventListener("mouseout", (ev) => {
             if (!mouse_disabled) {
                 pointerOut(ev);
             } else {
@@ -479,7 +523,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             }
             return false;
         });
-        div.addEventListener("contextmenu", (ev) => {
+        canvas.addEventListener("contextmenu", (ev) => {
             if (!mouse_disabled) {
                 pointerUp(ev, false);
             } else {
@@ -487,7 +531,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             }
             return false;
         });
-        div.addEventListener("focus", (ev) => {
+        canvas.addEventListener("focus", (ev) => {
             ev.preventDefault();
             return false;
         });
@@ -505,9 +549,9 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 mouse_disabled = setTimeout(() => {
                     mouse_disabled = 0;
                 }, 5000);
-                getRelativeEventPosition(ev, this.parent); // enables tracking of last ev position so on touch end can always tell where we released from
+                getRelativeEventPosition(ev); // enables tracking of last ev position so on touch end can always tell where we released from
 
-                if (ev.target === div) {
+                if (ev.target === canvas) {
                     lastX = ev.touches[0].clientX;
                     lastY = ev.touches[0].clientY;
                     startX = ev.touches[0].clientX;
@@ -525,7 +569,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 // Stop a touch screen device always auto scrolling to the chat input box if it is active when you make a move
                 const currentElement = document.activeElement;
                 if (
-                    ev.target === div &&
+                    ev.target === canvas &&
                     currentElement &&
                     currentElement instanceof HTMLElement &&
                     currentElement.tagName.toLowerCase() === "input"
@@ -540,7 +584,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                     mouse_disabled = 0;
                 }, 5000);
 
-                if (ev.target === div) {
+                if (ev.target === canvas) {
                     if (
                         Math.sqrt(
                             (startX - lastX) * (startX - lastX) +
@@ -566,9 +610,9 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 mouse_disabled = setTimeout(() => {
                     mouse_disabled = 0;
                 }, 5000);
-                getRelativeEventPosition(ev, this.parent); // enables tracking of last ev position so on touch end can always tell where we released from
+                getRelativeEventPosition(ev); // enables tracking of last ev position so on touch end can always tell where we released from
 
-                if (ev.target === div) {
+                if (ev.target === canvas) {
                     lastX = ev.touches[0].clientX;
                     lastY = ev.touches[0].clientY;
                     if (this.mode === "analyze" && this.analyze_tool === "draw") {
@@ -600,13 +644,8 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
     }
     public clearAnalysisDrawing(): void {
         this.pen_marks = [];
-        /*
         if (this.pen_ctx) {
             this.pen_ctx.clearRect(0, 0, this.metrics.width, this.metrics.height);
-        }
-        */
-        if (this.pen_layer) {
-            this.pen_layer.innerHTML = "";
         }
     }
     private xy2pen(x: number, y: number): [number, number] {
@@ -624,7 +663,6 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         return [(x / 64 - lx) * this.square_size, (y / 64 - ly) * this.square_size];
     }
     private setPenStyle(color: string): void {
-        /*
         if (!this.pen_ctx) {
             throw new Error(`setPenStyle called with null pen_ctx`);
         }
@@ -632,12 +670,11 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         this.pen_ctx.strokeStyle = color;
         this.pen_ctx.lineWidth = Math.max(1, Math.round(this.square_size * 0.1));
         this.pen_ctx.lineCap = "round";
-        */
     }
     private onPenStart(ev: MouseEvent | TouchEvent): void {
-        this.attachPenLayer();
+        this.attachPenCanvas();
 
-        const pos = getRelativeEventPosition(ev, this.parent);
+        const pos = getRelativeEventPosition(ev);
         this.last_pen_position = this.xy2pen(pos.x, pos.y);
         this.current_pen_mark = { color: this.analyze_subtool, points: this.xy2pen(pos.x, pos.y) };
         this.pen_marks.push(this.current_pen_mark);
@@ -646,11 +683,14 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         this.syncReviewMove({ pen: this.analyze_subtool, pp: this.xy2pen(pos.x, pos.y) });
     }
     private onPenMove(ev: MouseEvent | TouchEvent): void {
+        if (!this.pen_ctx) {
+            throw new Error(`onPenMove called with null pen_ctx`);
+        }
         if (!this.last_pen_position || !this.current_pen_mark) {
             throw new Error(`onPenMove called with invalid last pen position or current pen mark`);
         }
 
-        const pos = getRelativeEventPosition(ev, this.parent);
+        const pos = getRelativeEventPosition(ev);
         const start = this.last_pen_position;
         const s = this.pen2xy(start[0], start[1]);
         const end = this.xy2pen(pos.x, pos.y);
@@ -667,20 +707,10 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         this.current_pen_mark.points.push(dx);
         this.current_pen_mark.points.push(dy);
 
-        /*
         this.pen_ctx.beginPath();
         this.pen_ctx.moveTo(s[0], s[1]);
         this.pen_ctx.lineTo(e[0], e[1]);
         this.pen_ctx.stroke();
-        */
-
-        const path = `M ${s[0]} ${s[1]} L ${e[0]} ${e[1]}`;
-        const path_element = document.createElementNS("http://www.w3.org/2000/svg", "path");
-        path_element.setAttribute("d", path);
-        path_element.setAttribute("stroke", this.analyze_subtool);
-        path_element.setAttribute("stroke-width", "3");
-        path_element.setAttribute("fill", "none");
-        this.pen_layer?.appendChild(path_element);
 
         this.syncReviewMove({ pp: [dx, dy] });
     }
@@ -691,30 +721,28 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         if (!pen_marks.length) {
             return;
         }
-        this.attachPenLayer();
+        this.attachPenCanvas();
+        if (!this.pen_ctx) {
+            throw new Error(`onPenMove called with null pen_ctx`);
+        }
         this.clearAnalysisDrawing();
         this.pen_marks = pen_marks;
-
         for (let i = 0; i < pen_marks.length; ++i) {
             const stroke = pen_marks[i];
             this.setPenStyle(stroke.color);
 
             let px = stroke.points[0];
             let py = stroke.points[1];
+            this.pen_ctx.beginPath();
             const pt = this.pen2xy(px, py);
-            let path = `M ${pt[0]} ${pt[1]}`;
+            this.pen_ctx.moveTo(pt[0], pt[1]);
             for (let j = 2; j < stroke.points.length; j += 2) {
                 px += stroke.points[j];
                 py += stroke.points[j + 1];
                 const pt = this.pen2xy(px, py);
-                path += ` L ${pt[0]} ${pt[1]}`;
+                this.pen_ctx.lineTo(pt[0], pt[1]);
             }
-            const path_element = document.createElementNS("http://www.w3.org/2000/svg", "path");
-            path_element.setAttribute("d", path);
-            path_element.setAttribute("stroke", stroke.color);
-            path_element.setAttribute("stroke-width", "3");
-            path_element.setAttribute("fill", "none");
-            this.pen_layer?.appendChild(path_element);
+            this.pen_ctx.stroke();
         }
     }
     private onTap(
@@ -735,13 +763,13 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             return;
         }
 
-        // If there are modes where right click should not behave as if you clicked a placement on the svg
+        // If there are modes where right click should not behave as if you clicked a placement on the canvas
         // then return here instead of proceeding.
         if (right_click && this.mode === "play") {
             return;
         }
 
-        const pos = getRelativeEventPosition(event, this.parent);
+        const pos = getRelativeEventPosition(event);
         const xx = pos.x;
         const yy = pos.y;
 
@@ -1140,7 +1168,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             return;
         }
 
-        const pos = getRelativeEventPosition(event, this.parent);
+        const pos = getRelativeEventPosition(event);
         const pt = this.xy2ij(pos.x, pos.y);
 
         if (this.__last_pt.i === pt.i && this.__last_pt.j === pt.j) {
@@ -1198,41 +1226,25 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         if (!this.drawing_enabled || this.no_display) {
             return;
         }
+        const ctx = this.ctx;
+        if (!ctx) {
+            return;
+        }
         if (i < 0 || j < 0) {
             return;
         }
-
-        let cell = this.grid[j][i];
-        const shadow_cell = this.shadow_grid[j][i];
-
-        if (cell) {
-            cell.remove();
-        }
-        if (shadow_cell) {
-            shadow_cell.remove();
-        }
-
-        cell = this.grid[j][i] = document.createElementNS("http://www.w3.org/2000/svg", "g");
-        this.grid_layer!.appendChild(cell);
-
-        const ss = this.square_size;
-        let ox = this.draw_left_labels ? ss : 0;
-        let oy = this.draw_top_labels ? ss : 0;
+        const s = this.square_size;
+        let ox = this.draw_left_labels ? s : 0;
+        let oy = this.draw_top_labels ? s : 0;
         if (this.bounds.left > 0) {
-            ox = -ss * this.bounds.left;
+            ox = -s * this.bounds.left;
         }
         if (this.bounds.top > 0) {
-            oy = -ss * this.bounds.top;
+            oy = -s * this.bounds.top;
         }
 
-        const l = i * ss + ox;
-        const r = (i + 1) * ss + ox;
-        const t = j * ss + oy;
-        const b = (j + 1) * ss + oy;
-
-        const cx = l + this.metrics.mid;
-        const cy = t + this.metrics.mid;
-
+        let cx: number;
+        let cy: number;
         let draw_last_move = !this.dont_draw_last_move;
 
         let stone_color = 0;
@@ -1295,9 +1307,42 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             have_text_to_draw = true;
         }
 
-        /* Fade our lines if we have text to draw */
-        if (have_text_to_draw) {
-            /* draw lighter colored lines */
+        /* clear and draw lines */
+        {
+            const l = i * s + ox;
+            const r = (i + 1) * s + ox;
+            const t = j * s + oy;
+            const b = (j + 1) * s + oy;
+
+            ctx.clearRect(l, t, r - l, b - t);
+            if (this.shadow_ctx) {
+                let shadow_offset = this.square_size * 0.1;
+                this.shadow_ctx.clearRect(
+                    l + shadow_offset,
+                    t + shadow_offset,
+                    this.square_size,
+                    this.square_size,
+                );
+                shadow_offset = this.square_size * 0.2;
+                this.shadow_ctx.clearRect(
+                    l + shadow_offset,
+                    t + shadow_offset,
+                    this.square_size,
+                    this.square_size,
+                );
+                shadow_offset = this.square_size * 0.3;
+                this.shadow_ctx.clearRect(
+                    l + shadow_offset,
+                    t + shadow_offset,
+                    this.square_size,
+                    this.square_size,
+                );
+            }
+
+            cx = l + this.metrics.mid;
+            cy = t + this.metrics.mid;
+
+            /* draw line */
             let sx = l;
             let ex = r;
             const mx = (r + l) / 2 - this.metrics.offset;
@@ -1325,20 +1370,27 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 }
             }
 
-            const path = document.createElementNS("http://www.w3.org/2000/svg", "path");
-            path.setAttribute("stroke", this.theme_faded_line_color);
-            path.setAttribute("stroke-width", this.square_size < 5 ? "0.2" : "1");
-            path.setAttribute("fill", "none");
-            path.setAttribute(
-                "d",
-                `
-                M ${Math.floor(sx)} ${my} L ${Math.floor(ex)} ${my}
-                M ${mx} ${Math.floor(sy)} L ${mx} ${Math.floor(ey)} 
-            `,
-            );
-            cell.appendChild(path);
+            if (this.square_size < 5) {
+                ctx.lineWidth = 0.2;
+            } else {
+                ctx.lineWidth = 1;
+            }
+            if (have_text_to_draw) {
+                ctx.strokeStyle = this.theme_faded_line_color;
+            } else {
+                ctx.strokeStyle = this.theme_line_color;
+            }
+            ctx.lineCap = "butt";
+            ctx.beginPath();
+            ctx.moveTo(Math.floor(sx), my);
+            ctx.lineTo(Math.floor(ex), my);
+            ctx.moveTo(mx, Math.floor(sy));
+            ctx.lineTo(mx, Math.floor(ey));
+            ctx.stroke();
+        }
 
-            /* Draw star points */
+        /* Draw star points */
+        {
             let star_radius;
             if (this.square_size < 5) {
                 star_radius = 0.5;
@@ -1377,28 +1429,41 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             }
 
             if (draw_star_point) {
-                const circ = document.createElementNS("http://www.w3.org/2000/svg", "circle");
-                circ.setAttribute("cx", cx.toString());
-                circ.setAttribute("cy", cy.toString());
-                circ.setAttribute("r", star_radius.toString());
-                circ.setAttribute("fill", this.theme_faded_star_color);
-                cell.appendChild(circ);
+                ctx.beginPath();
+                ctx.fillStyle = this.theme_star_color;
+                if (have_text_to_draw) {
+                    ctx.fillStyle = this.theme_faded_star_color;
+                }
+                ctx.arc(
+                    cx,
+                    cy,
+                    star_radius,
+                    0.001,
+                    2 * Math.PI,
+                    false,
+                ); /* 0.001 to workaround fucked up chrome 27 bug */
+                ctx.fill();
             }
         }
 
         /* Heatmap */
+
         if (this.heatmap) {
             if (this.heatmap[j][i] > 0.001) {
                 const color = "#00FF00";
+                ctx.lineCap = "square";
+                ctx.save();
+                ctx.beginPath();
+                ctx.globalAlpha = Math.min(this.heatmap[j][i], 0.5);
                 const r = Math.floor(this.square_size * 0.5) - 0.5;
-                const rect = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-                rect.setAttribute("x", (cx - r).toFixed(1));
-                rect.setAttribute("y", (cy - r).toFixed(1));
-                rect.setAttribute("width", (r * 2).toFixed(1));
-                rect.setAttribute("height", (r * 2).toFixed(1));
-                rect.setAttribute("fill-opacity", Math.min(this.heatmap[j][i], 0.5).toFixed(2));
-                rect.setAttribute("fill", color);
-                cell.appendChild(rect);
+                ctx.moveTo(cx - r, cy - r);
+                ctx.lineTo(cx + r, cy - r);
+                ctx.lineTo(cx + r, cy + r);
+                ctx.lineTo(cx - r, cy + r);
+                ctx.lineTo(cx - r, cy - r);
+                ctx.fillStyle = color;
+                ctx.fill();
+                ctx.restore();
             }
         }
 
@@ -1410,43 +1475,57 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 pos.color
             ) {
                 const color = pos.color ? pos.color : pos.hint ? "#8EFF0A" : "#FF8E0A";
+
+                ctx.lineCap = "square";
+                ctx.save();
+                ctx.beginPath();
+                ctx.globalAlpha = 0.6;
                 const r = Math.floor(this.square_size * 0.5) - 0.5;
-                const rect = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-                rect.setAttribute("x", (cx - r).toFixed(1));
-                rect.setAttribute("y", (cy - r).toFixed(1));
-                rect.setAttribute("width", (r * 2).toFixed(1));
-                rect.setAttribute("height", (r * 2).toFixed(1));
-                rect.setAttribute("fill-opacity", "0.6");
-                rect.setAttribute("fill", color);
-                cell.appendChild(rect);
+                ctx.moveTo(cx - r, cy - r);
+                ctx.lineTo(cx + r, cy - r);
+                ctx.lineTo(cx + r, cy + r);
+                ctx.lineTo(cx - r, cy + r);
+                ctx.lineTo(cx - r, cy - r);
+                ctx.fillStyle = color;
+                ctx.fill();
+                ctx.restore();
             }
         }
 
         /* Colored stones */
+
         if (this.colored_circles) {
             if (this.colored_circles[j][i]) {
                 const circle = this.colored_circles[j][i];
+                const color = circle.color;
+
+                ctx.save();
+                ctx.globalAlpha = 1.0;
                 const radius = Math.floor(this.square_size * 0.5) - 0.5;
                 let lineWidth = radius * (circle.border_width || 0.1);
+
                 if (lineWidth < 0.3) {
                     lineWidth = 0;
                 }
-
-                const circ = document.createElementNS("http://www.w3.org/2000/svg", "circle");
-                circ.setAttribute("class", "colored-circle");
-                circ.setAttribute("fill", circle.color);
-                if (circle.border_color) {
-                    circ.setAttribute("stroke", circle.border_color);
-                }
+                ctx.fillStyle = color;
+                ctx.strokeStyle = circle.border_color || "#000000";
                 if (lineWidth > 0) {
-                    circ.setAttribute("stroke-width", lineWidth.toFixed(1));
-                } else {
-                    circ.setAttribute("stroke-width", "1px");
+                    ctx.lineWidth = lineWidth;
                 }
-                circ.setAttribute("cx", cx.toString());
-                circ.setAttribute("cy", cy.toString());
-                circ.setAttribute("r", Math.max(0.1, radius - lineWidth / 2).toString());
-                cell.appendChild(circ);
+                ctx.beginPath();
+                ctx.arc(
+                    cx,
+                    cy,
+                    Math.max(0.1, radius - lineWidth / 2),
+                    0.001,
+                    2 * Math.PI,
+                    false,
+                ); /* 0.001 to workaround fucked up chrome bug */
+                if (lineWidth > 0) {
+                    ctx.stroke();
+                }
+                ctx.fill();
+                ctx.restore();
             }
         }
 
@@ -1573,8 +1652,15 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                         return;
                     }
 
-                    const stone_transparent = translucent || !stone_color;
-
+                    ctx.save();
+                    let shadow_ctx: CanvasRenderingContext2D | null | undefined = this.shadow_ctx;
+                    if (!stone_color || translucent) {
+                        ctx.globalAlpha = stoneAlphaValue;
+                        shadow_ctx = null;
+                    }
+                    if (shadow_ctx === undefined) {
+                        shadow_ctx = null;
+                    }
                     if (color === 1) {
                         const stone = this.theme_black.getStone(
                             i,
@@ -1582,18 +1668,14 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                             this.theme_black_stones,
                             this,
                         );
-                        const [elt, shadow] = this.theme_black.placeBlackStoneSVG(
-                            cell,
-                            stone_transparent ? undefined : this.shadow_layer,
+                        this.theme_black.placeBlackStone(
+                            ctx,
+                            shadow_ctx,
                             stone,
                             cx,
                             cy,
                             this.theme_stone_radius,
                         );
-                        this.shadow_grid[j][i] = shadow;
-                        if (stone_transparent) {
-                            elt.setAttribute("opacity", stoneAlphaValue.toString());
-                        }
                     } else {
                         const stone = this.theme_white.getStone(
                             i,
@@ -1601,19 +1683,16 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                             this.theme_white_stones,
                             this,
                         );
-                        const [elt, shadow] = this.theme_white.placeWhiteStoneSVG(
-                            cell,
-                            stone_transparent ? undefined : this.shadow_layer,
+                        this.theme_white.placeWhiteStone(
+                            ctx,
+                            shadow_ctx,
                             stone,
                             cx,
                             cy,
                             this.theme_stone_radius,
                         );
-                        this.shadow_grid[j][i] = shadow;
-                        if (stone_transparent) {
-                            elt.setAttribute("opacity", stoneAlphaValue.toString());
-                        }
                     }
+                    ctx.restore();
                 }
 
                 if (
@@ -1623,26 +1702,32 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                     this.colored_circles[j][i]
                 ) {
                     const circle = this.colored_circles[j][i];
+
+                    ctx.save();
+                    ctx.globalAlpha = 1.0;
                     const radius = Math.floor(this.square_size * 0.5) - 0.5;
                     let lineWidth = radius * (circle.border_width || 0.1);
+
                     if (lineWidth < 0.3) {
                         lineWidth = 0;
                     }
-
-                    const circ = document.createElementNS("http://www.w3.org/2000/svg", "circle");
-                    circ.setAttribute("class", "colored-circle");
-                    if (circle.border_color) {
-                        circ.setAttribute("stroke", circle.border_color || "#000000");
-                    }
+                    ctx.strokeStyle = circle.border_color || "#000000";
                     if (lineWidth > 0) {
-                        circ.setAttribute("stroke-width", `${lineWidth.toFixed(1)}px`);
-                    } else {
-                        circ.setAttribute("stroke-width", "1px");
+                        ctx.lineWidth = lineWidth;
                     }
-                    circ.setAttribute("cx", cx.toString());
-                    circ.setAttribute("cy", cy.toString());
-                    circ.setAttribute("r", Math.max(0.1, radius - lineWidth / 2).toString());
-                    cell.appendChild(circ);
+                    ctx.beginPath();
+                    ctx.arc(
+                        cx,
+                        cy,
+                        Math.max(0.1, radius - lineWidth / 2),
+                        0.001,
+                        2 * Math.PI,
+                        false,
+                    ); /* 0.001 to workaround fucked up chrome bug */
+                    if (lineWidth > 0) {
+                        ctx.stroke();
+                    }
+                    ctx.restore();
                 }
 
                 /* Red X if the stone is marked for removal */
@@ -1656,14 +1741,12 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                         this.score_estimator &&
                         this.score_estimator.board[j][i] &&
                         this.score_estimator.removal[j][i]) ||
-                    //(this.mode === "analyze" && pos.stone_removed)
                     pos.stone_removed
                 ) {
                     draw_red_x = true;
                 }
             }
         }
-
         if (
             draw_red_x ||
             (this.mode === "analyze" &&
@@ -1678,25 +1761,21 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 this.last_hover_square.x === i &&
                 this.last_hover_square.y === j)
         ) {
-            const r = Math.max(1, this.metrics.mid * 0.75);
-            const cross = document.createElementNS("http://www.w3.org/2000/svg", "path");
-            cross.setAttribute("class", "removal-cross");
-            cross.setAttribute("stroke", "#ff0000");
-            cross.setAttribute("stroke-width", `${this.square_size * 0.125}px`);
-            cross.setAttribute("fill", "none");
-            cross.setAttribute(
-                "d",
-                `
-                    M ${cx - r} ${cy - r}
-                    L ${cx + r} ${cy + r}
-                    M ${cx + r} ${cy - r}
-                    L ${cx - r} ${cy + r}
-                `,
-            );
             const opacity = this.engine.board[j][i] ? 1.0 : 0.2;
-            cross.setAttribute("stroke-opacity", opacity?.toString());
-
-            cell.appendChild(cross);
+            ctx.lineCap = "square";
+            ctx.save();
+            ctx.beginPath();
+            ctx.lineWidth = this.square_size * 0.125;
+            ctx.globalAlpha = opacity;
+            const r = Math.max(1, this.metrics.mid * 0.65);
+            ctx.moveTo(cx - r, cy - r);
+            ctx.lineTo(cx + r, cy + r);
+            ctx.moveTo(cx + r, cy - r);
+            ctx.lineTo(cx - r, cy + r);
+            ctx.strokeStyle = "#ff0000";
+            ctx.stroke();
+            ctx.restore();
+            draw_last_move = false;
         }
 
         /* Draw Scores */
@@ -1718,8 +1797,9 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                     this.last_hover_square.x === i &&
                     this.last_hover_square.y === j)
             ) {
-                let color = pos.score;
+                ctx.beginPath();
 
+                let color = pos.score;
                 if (
                     this.scoring_mode &&
                     this.score_estimator &&
@@ -1759,38 +1839,31 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                     color = this.analyze_subtool;
                 }
 
-                const r = this.square_size * 0.15;
-                const rect = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-                rect.setAttribute("x", (cx - r).toFixed(1));
-                rect.setAttribute("y", (cy - r).toFixed(1));
-                rect.setAttribute("width", (r * 2).toFixed(1));
-                rect.setAttribute("height", (r * 2).toFixed(1));
                 if (color === "white") {
-                    rect.setAttribute("fill", this.theme_black_text_color);
-                    rect.setAttribute("stroke", "#777777");
-                }
-                if (color === "black") {
-                    rect.setAttribute("fill", this.theme_white_text_color);
-                    rect.setAttribute("stroke", "#888888");
-                }
-                if (color === "dame") {
-                    rect.setAttribute("fill-opacity", "0.2");
-                    rect.setAttribute("stroke", "#365FE6");
-                }
-                if (color === "seal") {
-                    rect.setAttribute("fill-opacity", "0.8");
-                    rect.setAttribute("fill", "#ff4444");
-                    rect.setAttribute("stroke", "#E079CE");
+                    ctx.fillStyle = this.theme_black_text_color;
+                    ctx.strokeStyle = "#777777";
+                } else if (color === "black") {
+                    ctx.fillStyle = this.theme_white_text_color;
+                    ctx.strokeStyle = "#888888";
+                } else if (color === "dame") {
+                    ctx.fillStyle = "#ff0000";
+                    ctx.strokeStyle = "#365FE6";
+                } else if (color === "seal") {
+                    ctx.fillStyle = "#ff0000";
+                    ctx.strokeStyle = "#E079CE";
                 }
                 if (color?.[0] === "#") {
-                    rect.setAttribute("fill", color);
-                    rect.setAttribute("stroke", color_blend("#888888", color));
+                    ctx.fillStyle = color;
+                    ctx.strokeStyle = color_blend("#888888", color);
                 }
-                rect.setAttribute(
-                    "stroke-width",
-                    (Math.ceil(this.square_size * 0.065) - 0.5).toFixed(1),
-                );
-                cell.appendChild(rect);
+                ctx.lineWidth = Math.ceil(this.square_size * 0.065) - 0.5;
+
+                const r = this.square_size * 0.15;
+                ctx.rect(cx - r, cy - r, r * 2, r * 2);
+                if (color !== "dame") {
+                    ctx.fill();
+                }
+                ctx.stroke();
             }
         }
 
@@ -1858,54 +1931,70 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
             if (letter) {
                 letter_was_drawn = true;
-                const text = document.createElementNS("http://www.w3.org/2000/svg", "text");
-                text.setAttribute("class", "letter");
-                text.setAttribute("fill", text_color);
-                text.setAttribute(
-                    "font-size",
-                    subscript ? `${this.square_size * 0.4}px` : `${this.square_size * 0.5}px`,
+                ctx.save();
+                ctx.fillStyle = text_color;
+                const [, , metrics] = fitText(
+                    ctx,
+                    letter,
+                    `bold FONT_SIZEpx ${GOBAN_FONT}`,
+                    this.square_size * 0.4,
+                    this.square_size * 0.8 * (subscript ? 0.9 : 1.0),
                 );
-                text.setAttribute("text-anchor", "middle");
-                text.setAttribute("x", cx.toString());
-                let yy = cy;
-                yy += this.square_size / 6;
+
+                const xx = cx - metrics.width / 2;
+                let yy =
+                    cy +
+                    (/WebKit|Trident/.test(navigator.userAgent)
+                        ? this.square_size * -0.03
+                        : 1); /* middle centering is different on firefox */
+
                 if (subscript) {
                     yy -= this.square_size * 0.15;
                 }
-                text.setAttribute("y", yy.toString());
-                text.textContent = letter;
+
+                ctx.textBaseline = "middle";
                 if (transparent) {
-                    text.setAttribute("fill-opacity", "0.6");
+                    ctx.globalAlpha = 0.6;
                 }
-                cell.appendChild(text);
+                ctx.fillText(letter, xx, yy);
                 draw_last_move = false;
+                ctx.restore();
             }
 
             if (subscript) {
                 letter_was_drawn = true;
-                const text = document.createElementNS("http://www.w3.org/2000/svg", "text");
-                text.setAttribute("class", "subscript");
-                text.setAttribute("fill", text_color);
-                text.setAttribute("font-size", `${this.square_size * 0.4}px`);
-                text.setAttribute("text-anchor", "middle");
-                text.setAttribute("x", cx.toString());
-                let yy = cy;
-                yy -= this.square_size / 6;
+                ctx.save();
+                ctx.fillStyle = text_color;
+                if (letter && subscript === "0") {
+                    subscript = "0.0"; // clarifies the markings on the blue move typically
+                }
+
+                const [, , metrics] = fitText(
+                    ctx,
+                    subscript,
+                    `bold FONT_SIZEpx ${GOBAN_FONT}`,
+                    this.square_size * 0.4,
+                    this.square_size * 0.8 * (letter ? 0.9 : 1.0),
+                );
+
+                const xx = cx - metrics.width / 2;
+                let yy =
+                    cy +
+                    (/WebKit|Trident/.test(navigator.userAgent)
+                        ? this.square_size * -0.03
+                        : 1); /* middle centering is different on firefox */
+
                 if (letter) {
-                    yy += this.square_size * 0.6;
-                } else {
-                    yy += this.square_size * 0.31;
+                    yy += this.square_size * 0.3;
                 }
-                if (pos.sub_triangle) {
-                    yy -= this.square_size * 0.08;
-                }
-                text.setAttribute("y", yy.toString());
-                text.textContent = subscript;
+
+                ctx.textBaseline = "middle";
                 if (transparent) {
-                    text.setAttribute("fill-opacity", "0.6");
+                    ctx.globalAlpha = 0.6;
                 }
-                cell.appendChild(text);
+                ctx.fillText(subscript, xx, yy);
                 draw_last_move = false;
+                ctx.restore();
             }
         }
 
@@ -1938,21 +2027,18 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             }
 
             if (pos.circle || hover_mark === "circle") {
-                const circ = document.createElementNS("http://www.w3.org/2000/svg", "circle");
-                circ.setAttribute("class", "circle");
-                circ.setAttribute("fill", "none");
-                circ.setAttribute("stroke", symbol_color);
-                circ.setAttribute("stroke-width", `${this.square_size * 0.075}px`);
-                circ.setAttribute("cx", cx.toString());
-                circ.setAttribute("cy", cy.toString());
-                circ.setAttribute(
-                    "r",
-                    Math.max(0.1, this.square_size * this.circle_radius).toFixed(2),
-                );
+                ctx.lineCap = "round";
+                ctx.save();
+                ctx.beginPath();
                 if (transparent) {
-                    circ.setAttribute("stroke-opacity", "0.6");
+                    ctx.globalAlpha = 0.6;
                 }
-                cell.appendChild(circ);
+                ctx.strokeStyle = symbol_color;
+                ctx.lineWidth = this.square_size * 0.075;
+                const r = Math.max(0.1, this.square_size * this.circle_radius);
+                ctx.arc(cx, cy, r, 0, 2 * Math.PI, false);
+                ctx.stroke();
+                ctx.restore();
                 draw_last_move = false;
             }
             if (
@@ -1964,80 +2050,72 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             ) {
                 let scale = 1.0;
                 let oy = 0.0;
-                let line_width = this.square_size * 0.075;
                 if (pos.sub_triangle) {
                     scale = 0.5;
                     oy = this.square_size * 0.3;
                     transparent = false;
-                    line_width *= 0.5;
                 }
-
-                const triangle = document.createElementNS("http://www.w3.org/2000/svg", "polygon");
-                triangle.setAttribute("class", "triangle");
-                triangle.setAttribute("fill", "none");
-                triangle.setAttribute("stroke", symbol_color);
-                triangle.setAttribute("stroke-width", `${line_width}px`);
-                const r = this.square_size * 0.3 * scale;
-                let theta = -(Math.PI * 2) / 4;
-                const points = [];
-                points.push([cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta)]);
-                theta += (Math.PI * 2) / 3;
-                points.push([cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta)]);
-                theta += (Math.PI * 2) / 3;
-                points.push([cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta)]);
-                theta += (Math.PI * 2) / 3;
-                points.push([cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta)]);
-                triangle.setAttribute(
-                    "points",
-                    points.map((p) => `${p[0].toFixed(2)},${p[1].toFixed(2)}`).join(" "),
-                );
+                ctx.lineCap = "round";
+                ctx.save();
+                ctx.beginPath();
                 if (transparent) {
-                    triangle.setAttribute("stroke-opacity", "0.6");
+                    ctx.globalAlpha = 0.6;
                 }
-                cell.appendChild(triangle);
+                ctx.strokeStyle = symbol_color;
+                if (pos.chat_triangle) {
+                    ctx.strokeStyle = "#00aaFF";
+                }
+                ctx.lineWidth = this.square_size * 0.075 * scale;
+                let theta = -(Math.PI * 2) / 4;
+                const r = this.square_size * 0.3 * scale;
+                ctx.moveTo(cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta));
+                theta += (Math.PI * 2) / 3;
+                ctx.lineTo(cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta));
+                theta += (Math.PI * 2) / 3;
+                ctx.lineTo(cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta));
+                theta += (Math.PI * 2) / 3;
+                ctx.lineTo(cx + r * Math.cos(theta), cy + oy + r * Math.sin(theta));
+                ctx.stroke();
+                ctx.restore();
                 draw_last_move = false;
             }
             if (pos.cross || hover_mark === "cross") {
-                const r = Math.max(1, this.metrics.mid * 0.35);
-                const cross = document.createElementNS("http://www.w3.org/2000/svg", "path");
-                cross.setAttribute("class", "cross");
-                cross.setAttribute("stroke", symbol_color);
-                cross.setAttribute("stroke-width", `${this.square_size * 0.075}px`);
-                cross.setAttribute("fill", "none");
-                cross.setAttribute(
-                    "d",
-                    `
-                    M ${cx - r} ${cy - r}
-                    L ${cx + r} ${cy + r}
-                    M ${cx + r} ${cy - r}
-                    L ${cx - r} ${cy + r}
-                `,
-                );
+                ctx.lineCap = "square";
+                ctx.save();
+                ctx.beginPath();
+                ctx.lineWidth = this.square_size * 0.075;
                 if (transparent) {
-                    cross.setAttribute("stroke-opacity", "0.6");
+                    ctx.globalAlpha = 0.6;
                 }
-
-                cell.appendChild(cross);
-
+                const r = Math.max(1, this.metrics.mid * 0.35);
+                ctx.moveTo(cx - r, cy - r);
+                ctx.lineTo(cx + r, cy + r);
+                ctx.moveTo(cx + r, cy - r);
+                ctx.lineTo(cx - r, cy + r);
+                ctx.strokeStyle = symbol_color;
+                ctx.stroke();
+                ctx.restore();
                 draw_last_move = false;
             }
 
             if (pos.square || hover_mark === "square") {
-                draw_last_move = false;
-                const square = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-                square.setAttribute("class", "square");
-                square.setAttribute("fill", "none");
-                square.setAttribute("stroke", symbol_color);
-                square.setAttribute("stroke-width", `${this.square_size * 0.075}px`);
-                const r = Math.max(1, this.metrics.mid * 0.4);
-                square.setAttribute("x", (cx - r).toFixed(2));
-                square.setAttribute("y", (cy - r).toFixed(2));
-                square.setAttribute("width", (r * 2).toFixed(2));
-                square.setAttribute("height", (r * 2).toFixed(2));
+                ctx.lineCap = "square";
+                ctx.save();
+                ctx.beginPath();
+                ctx.lineWidth = this.square_size * 0.075;
                 if (transparent) {
-                    square.setAttribute("stroke-opacity", "0.6");
+                    ctx.globalAlpha = 0.6;
                 }
-                cell.appendChild(square);
+                const r = Math.max(1, this.metrics.mid * 0.4);
+                ctx.moveTo(cx - r, cy - r);
+                ctx.lineTo(cx + r, cy - r);
+                ctx.lineTo(cx + r, cy + r);
+                ctx.lineTo(cx - r, cy + r);
+                ctx.lineTo(cx - r, cy - r);
+                ctx.strokeStyle = symbol_color;
+                ctx.stroke();
+                ctx.restore();
+                draw_last_move = false;
             }
         }
 
@@ -2065,28 +2143,20 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                             : this.theme_white_text_color;
 
                     if (this.submit_move) {
-                        draw_last_move = false;
-
+                        ctx.lineCap = "square";
+                        ctx.save();
+                        ctx.beginPath();
+                        ctx.lineWidth = this.square_size * 0.075;
+                        ctx.globalAlpha = this.last_move_opacity;
                         const r = Math.max(1, this.metrics.mid * 0.35) * 0.8;
-                        const cross = document.createElementNS(
-                            "http://www.w3.org/2000/svg",
-                            "path",
-                        );
-                        cross.setAttribute("class", "last-move");
-                        cross.setAttribute("stroke", color);
-                        cross.setAttribute("stroke-width", `${this.square_size * 0.075}px`);
-                        cross.setAttribute("fill", "none");
-                        cross.setAttribute("opacity", this.last_move_opacity.toString());
-                        cross.setAttribute(
-                            "d",
-                            `
-                            M ${cx - r} ${cy}
-                            L ${cx + r} ${cy}
-                            M ${cx} ${cy - r}
-                            L ${cx} ${cy + r}
-                        `,
-                        );
-                        cell.appendChild(cross);
+                        ctx.moveTo(cx - r, cy);
+                        ctx.lineTo(cx + r, cy);
+                        ctx.moveTo(cx, cy - r);
+                        ctx.lineTo(cx, cy + r);
+                        ctx.strokeStyle = color;
+                        ctx.stroke();
+                        ctx.restore();
+                        draw_last_move = false;
                     } else {
                         if (
                             this.engine.undo_requested &&
@@ -2094,42 +2164,34 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                             this.engine.undo_requested === this.engine.cur_move.move_number
                         ) {
                             const letter = "?";
+                            ctx.save();
+                            ctx.fillStyle = color;
+                            const metrics = ctx.measureText(letter);
+                            const xx = cx - metrics.width / 2;
+                            const yy =
+                                cy +
+                                (/WebKit|Trident/.test(navigator.userAgent)
+                                    ? this.square_size * -0.03
+                                    : 1); /* middle centering is different on firefox */
+                            ctx.textBaseline = "middle";
+                            ctx.fillText(letter, xx, yy);
                             draw_last_move = false;
-
-                            const text = document.createElementNS(
-                                "http://www.w3.org/2000/svg",
-                                "text",
-                            );
-                            text.setAttribute("class", "letter");
-                            text.setAttribute("fill", color);
-                            text.setAttribute("font-size", `${this.square_size * 0.5}px`);
-                            text.setAttribute("text-anchor", "middle");
-                            text.setAttribute("x", cx.toString());
-                            let yy = cy;
-                            yy += this.square_size / 6;
-                            text.setAttribute("y", yy.toString());
-                            text.textContent = letter;
-                            cell.appendChild(text);
+                            ctx.restore();
                         } else {
-                            const circ = document.createElementNS(
-                                "http://www.w3.org/2000/svg",
-                                "circle",
-                            );
+                            ctx.save();
+                            ctx.beginPath();
+                            ctx.strokeStyle = color;
+                            ctx.lineWidth = this.square_size * 0.075;
+                            ctx.globalAlpha = this.last_move_opacity;
                             let r = this.square_size * this.last_move_radius;
                             if (this.submit_move) {
                                 r = this.square_size * 0.3;
                             }
 
                             r = Math.max(0.1, r);
-                            circ.setAttribute("class", "last-move");
-                            circ.setAttribute("fill", "none");
-                            circ.setAttribute("stroke", color);
-                            circ.setAttribute("stroke-width", `${this.square_size * 0.075}px`);
-                            circ.setAttribute("cx", cx.toString());
-                            circ.setAttribute("cy", cy.toString());
-                            circ.setAttribute("opacity", this.last_move_opacity.toString());
-                            circ.setAttribute("r", r.toString());
-                            cell.appendChild(circ);
+                            ctx.arc(cx, cy, r, 0, 2 * Math.PI, false);
+                            ctx.stroke();
+                            ctx.restore();
                         }
                     }
                 }
@@ -2148,35 +2210,30 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                     ? this.stalling_score_estimate
                     : this.score_estimator;
             const est = se!.ownership[j][i];
+
+            ctx.beginPath();
+
             const color = est < 0 ? "white" : "black";
             const color_num = color === "black" ? 1 : 2;
 
             if (color_num !== stone_color) {
-                const r = this.square_size * 0.2 * Math.abs(est);
-                const rect = document.createElementNS("http://www.w3.org/2000/svg", "rect");
-                rect.setAttribute("x", (cx - r).toFixed(1));
-                rect.setAttribute("y", (cy - r).toFixed(1));
-                rect.setAttribute("width", (r * 2).toFixed(1));
-                rect.setAttribute("height", (r * 2).toFixed(1));
                 if (color === "white") {
-                    rect.setAttribute("fill", this.theme_black_text_color);
-                    rect.setAttribute("stroke", "#777777");
+                    ctx.fillStyle = this.theme_black_text_color;
+                    ctx.strokeStyle = "#777777";
+                } else if (color === "black") {
+                    ctx.fillStyle = this.theme_white_text_color;
+                    ctx.strokeStyle = "#888888";
                 }
-                if (color === "black") {
-                    rect.setAttribute("fill", this.theme_white_text_color);
-                    rect.setAttribute("stroke", "#888888");
-                }
-                rect.setAttribute(
-                    "stroke-width",
-                    (Math.ceil(this.square_size * 0.035) - 0.5).toFixed(1),
-                );
-                cell.appendChild(rect);
+                ctx.lineWidth = Math.ceil(this.square_size * 0.035) - 0.5;
+                const r = this.square_size * 0.2 * Math.abs(est);
+                ctx.rect(cx - r, cy - r, r * 2, r * 2);
+                ctx.fill();
+                ctx.stroke();
             }
         }
 
         this.__draw_state[j][i] = this.drawingHash(i, j);
     }
-
     private drawingHash(i: number, j: number): string {
         if (this.no_display) {
             return "";
@@ -2324,15 +2381,6 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 }
 
                 if (
-                    pos.blue_move &&
-                    this.colored_circles &&
-                    this.colored_circles[j] &&
-                    this.colored_circles[j][i]
-                ) {
-                    ret += "blue";
-                }
-
-                if (
                     (this.engine &&
                         this.engine.phase === "stone removal" &&
                         this.engine.last_official_move === this.engine.cur_move &&
@@ -2473,6 +2521,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 ) {
                     color = this.analyze_subtool;
                 }
+
                 if (
                     this.scoring_mode &&
                     this.score_estimator &&
@@ -2603,301 +2652,6 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
         return ret;
     }
-
-    private drawLines(force_clear?: boolean): void {
-        if (force_clear) {
-            if (this.lines_layer) {
-                this.lines_layer.remove();
-                delete this.lines_layer;
-            }
-        }
-
-        if (!this.lines_layer) {
-            const ss = this.square_size;
-            let ox = this.draw_left_labels ? ss : 0;
-            let oy = this.draw_top_labels ? ss : 0;
-
-            if (this.bounds.left > 0) {
-                ox = -ss * this.bounds.left;
-            }
-            if (this.bounds.top > 0) {
-                oy = -ss * this.bounds.top;
-            }
-
-            // lines go through center of our stone grid
-            ox += Math.round(ss / 2);
-            oy += Math.round(ss / 2);
-
-            // Tiny square sizes, as in the ones used to display puzzle icons
-            const TINY_SQUARE_SIZE = 10;
-
-            // Compute a line width that is rounded to the nearest 0.5 so we
-            // get crisp lines
-            const line_width =
-                ss > TINY_SQUARE_SIZE
-                    ? Math.round(2 * Math.round(Math.max(1, ss * 0.02))) * 0.5
-                    : // for very small boards, like puzzle icons, have faint lines
-                      ss * 0.08;
-            ox -= line_width * 0.5;
-            oy -= line_width * 0.5;
-
-            // Round to half pixel offsets odd widths for crisp lines
-            ox = Math.round(ox * 2.0) * 0.5;
-            oy = Math.round(oy * 2.0) * 0.5;
-
-            this.lines_layer = document.createElementNS("http://www.w3.org/2000/svg", "g");
-            const lines_path = document.createElementNS("http://www.w3.org/2000/svg", "path");
-            let path_str = "";
-            for (let x = 0; x < this.width; ++x) {
-                path_str += `M ${ox + x * ss} ${oy} L ${ox + x * ss} ${
-                    oy + (this.height - 1) * ss
-                } `;
-            }
-            for (let y = 0; y < this.height; ++y) {
-                path_str += `M ${ox} ${oy + y * ss} L ${ox + (this.width - 1) * ss} ${
-                    oy + y * ss
-                } `;
-            }
-            lines_path.setAttribute("d", path_str);
-            lines_path.setAttribute("stroke", this.theme_line_color);
-            if (ss > TINY_SQUARE_SIZE) {
-                lines_path.setAttribute("stroke-width", `${line_width.toFixed(0)}px`);
-            } else {
-                lines_path.setAttribute("stroke-width", `${line_width.toFixed(1)}px`);
-            }
-            lines_path.setAttribute("stroke-linecap", "square");
-            this.lines_layer.appendChild(lines_path);
-
-            // Hoshi / star points
-            let hoshi = null;
-
-            if (this.width === 19 && this.height === 19) {
-                hoshi = [
-                    [3, 3],
-                    [3, 9],
-                    [3, 15],
-                    [9, 3],
-                    [9, 9],
-                    [9, 15],
-                    [15, 3],
-                    [15, 9],
-                    [15, 15],
-                ];
-            }
-
-            if (this.width === 13 && this.height === 13) {
-                hoshi = [
-                    [3, 3],
-                    [3, 9],
-                    [6, 6],
-                    [9, 3],
-                    [9, 9],
-                ];
-            }
-
-            if (this.width === 9 && this.height === 9) {
-                hoshi = [
-                    [2, 2],
-                    [2, 6],
-                    [4, 4],
-                    [6, 2],
-                    [6, 6],
-                ];
-            }
-
-            if (hoshi) {
-                const r = this.square_size * 0.075;
-                for (let i = 0; i < hoshi.length; ++i) {
-                    const [hx, hy] = hoshi[i];
-                    const circle = document.createElementNS("http://www.w3.org/2000/svg", "circle");
-                    circle.setAttribute("cx", (ox + hx * ss).toString());
-                    circle.setAttribute("cy", (oy + hy * ss).toString());
-                    circle.setAttribute("r", `${r.toFixed(1)}px`);
-                    circle.setAttribute("fill", this.theme_star_color);
-                    this.lines_layer.appendChild(circle);
-                }
-            }
-            this.svg.appendChild(this.lines_layer);
-        }
-    }
-
-    private drawCoordinateLabels(force_clear?: boolean): void {
-        if (force_clear) {
-            if (this.coordinate_labels_layer) {
-                this.coordinate_labels_layer.remove();
-                delete this.coordinate_labels_layer;
-            }
-        }
-
-        if (!this.coordinate_labels_layer) {
-            this.coordinate_labels_layer = document.createElementNS(
-                "http://www.w3.org/2000/svg",
-                "g",
-            );
-            this.coordinate_labels_layer.setAttribute("class", "coordinate-labels");
-
-            let text_size = Math.round(this.square_size * 0.5);
-            let bold_or_not = "bold";
-            if (this.getCoordinateDisplaySystem() === "1-1") {
-                text_size *= 0.7;
-                bold_or_not = "";
-
-                if (this.height > 20) {
-                    text_size *= 0.7;
-                }
-            }
-
-            const place = (ch: string, x: number, y: number): void => {
-                /* places centered (horizontally & vertically) text at x,y */
-                const ox = 0;
-                const oy = this.square_size / 6;
-
-                const text = document.createElementNS("http://www.w3.org/2000/svg", "text");
-                text.setAttribute("x", (x + ox).toFixed(0));
-                text.setAttribute("y", (y + oy).toFixed(0));
-                text.setAttribute("font-size", `${Math.round(text_size)}px`);
-                text.setAttribute("font-size", `${Math.round(text_size)}px`);
-                text.setAttribute("font-weight", bold_or_not);
-                text.setAttribute("fill", this.theme_board.getLabelTextColor());
-                text.textContent = ch;
-                this.coordinate_labels_layer!.appendChild(text);
-            };
-            const v_place = (ch: string, x: number, y: number): void => {
-                /* places centered (horizontally & vertically) text at x,y, with text going down vertically. */
-                for (let i = 0; i < ch.length; ++i) {
-                    //const xx = x - text_size / 2;
-                    const H = text_size;
-                    const xx = x;
-                    let yy = y + text_size / 2.5;
-
-                    if (ch.length === 2) {
-                        yy = yy - H * 0.5 + i * H;
-                    }
-                    if (ch.length === 3) {
-                        yy = yy - H * 1 + i * H + 0.5;
-                    }
-
-                    const text = document.createElementNS("http://www.w3.org/2000/svg", "text");
-                    text.setAttribute("x", xx.toFixed(0));
-                    text.setAttribute("y", yy.toFixed(1));
-                    text.setAttribute("font-size", `${Math.round(text_size)}px`);
-                    text.setAttribute("font-weight", bold_or_not);
-                    text.setAttribute("fill", this.theme_board.getLabelTextColor());
-                    text.textContent = ch[i];
-                    this.coordinate_labels_layer!.appendChild(text);
-                }
-            };
-
-            const drawHorizontal = (i: number, j: number): void => {
-                switch (this.getCoordinateDisplaySystem()) {
-                    case "A1":
-                        for (let c = 0; c < this.width; ++i, ++c) {
-                            const x =
-                                (i -
-                                    this.bounds.left -
-                                    (this.bounds.left > 0 ? +this.draw_left_labels : 0)) *
-                                    this.square_size +
-                                this.square_size / 2;
-                            const y = j * this.square_size + this.square_size / 2;
-                            place(GoMath.pretty_coor_num2ch(c), x, y);
-                        }
-                        break;
-                    case "1-1":
-                        for (let c = 0; c < this.width; ++i, ++c) {
-                            const x =
-                                (i -
-                                    this.bounds.left -
-                                    (this.bounds.left > 0 ? +this.draw_left_labels : 0)) *
-                                    this.square_size +
-                                this.square_size / 2;
-                            const y = j * this.square_size + this.square_size / 2;
-                            place("" + (c + 1), x, y);
-                        }
-                        break;
-                }
-            };
-
-            const drawVertical = (i: number, j: number): void => {
-                switch (this.getCoordinateDisplaySystem()) {
-                    case "A1":
-                        for (let c = 0; c < this.height; ++j, ++c) {
-                            const x = i * this.square_size + this.square_size / 2;
-                            const y =
-                                (j -
-                                    this.bounds.top -
-                                    (this.bounds.top > 0 ? +this.draw_top_labels : 0)) *
-                                    this.square_size +
-                                this.square_size / 2;
-                            place("" + (this.height - c), x, y);
-                        }
-                        break;
-                    case "1-1":
-                        const chinese_japanese_numbers = [
-                            "一",
-                            "二",
-                            "三",
-                            "四",
-                            "五",
-                            "六",
-                            "七",
-                            "八",
-                            "九",
-                            "十",
-                            "十一",
-                            "十二",
-                            "十三",
-                            "十四",
-                            "十五",
-                            "十六",
-                            "十七",
-                            "十八",
-                            "十九",
-                            "二十",
-                            "二十一",
-                            "二十二",
-                            "二十三",
-                            "二十四",
-                            "二十五",
-                        ];
-                        for (let c = 0; c < this.height; ++j, ++c) {
-                            const x = i * this.square_size + this.square_size / 2;
-                            const y =
-                                (j -
-                                    this.bounds.top -
-                                    (this.bounds.top > 0 ? +this.draw_top_labels : 0)) *
-                                    this.square_size +
-                                this.square_size / 2;
-                            v_place(chinese_japanese_numbers[c], x, y);
-                        }
-                        break;
-                }
-            };
-
-            if (this.draw_top_labels && this.bounds.top === 0) {
-                drawHorizontal(this.draw_left_labels ? 1 : 0, 0);
-            }
-            if (this.draw_bottom_labels && this.bounds.bottom === this.height - 1) {
-                drawHorizontal(
-                    this.draw_left_labels ? 1 : 0,
-                    +this.draw_top_labels + this.bounded_height,
-                );
-            }
-            if (this.draw_left_labels && this.bounds.left === 0) {
-                drawVertical(0, this.draw_top_labels ? 1 : 0);
-            }
-            if (this.draw_right_labels && this.bounds.right === this.width - 1) {
-                drawVertical(+this.draw_left_labels + this.bounded_width, +this.draw_top_labels);
-            }
-
-            this.svg.appendChild(this.coordinate_labels_layer);
-        }
-    }
-
-    protected computeThemeStoneRadius(): number {
-        const r = this.square_size * 0.5;
-        return Math.max(1, r);
-    }
-
     public redraw(force_clear?: boolean): void {
         if (!this.ready_to_draw) {
             return;
@@ -2924,13 +2678,43 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         ) {
             force_clear = true;
             try {
+                //this.parent.css({"width": metrics.width + "px", "height": metrics.height + "px"});
                 this.parent.style.width = metrics.width + "px";
                 this.parent.style.height = metrics.height + "px";
-                this.svg.setAttribute("width", metrics.width.toString());
-                this.svg.setAttribute("height", metrics.height.toString());
+                resizeDeviceScaledCanvas(this.board, metrics.width, metrics.height);
+
+                if (this.pen_layer) {
+                    if (this.pen_marks.length) {
+                        resizeDeviceScaledCanvas(this.pen_layer, metrics.width, metrics.height);
+                        const ctx = this.pen_layer.getContext("2d", { willReadFrequently: true });
+                        if (ctx) {
+                            this.pen_ctx = ctx;
+                        } else {
+                            throw new Error(`Failed to obtain drawing context for pen layer`);
+                        }
+                    } else {
+                        this.detachPenCanvas();
+                    }
+                }
+
+                if (this.shadow_layer) {
+                    resizeDeviceScaledCanvas(this.shadow_layer, metrics.width, metrics.height);
+                    const ctx = this.shadow_layer.getContext("2d", { willReadFrequently: true });
+                    if (ctx) {
+                        this.shadow_ctx = ctx;
+                    } else {
+                        throw new Error(`Failed to obtain drawing context for shadow layer`);
+                    }
+                }
 
                 this.__set_board_width = metrics.width;
                 this.__set_board_height = metrics.height;
+                const ctx = this.board.getContext("2d", { willReadFrequently: true });
+                if (ctx) {
+                    this.ctx = ctx;
+                } else {
+                    throw new Error(`Failed to obtain drawing context for board`);
+                }
 
                 this.setThemes(this.getSelectedThemes(), true);
             } catch (e) {
@@ -2940,37 +2724,161 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 return;
             }
         }
+        const ctx = this.ctx;
 
-        this.drawLines(force_clear);
-        this.drawCoordinateLabels(force_clear);
+        const place = (ch: string, x: number, y: number): void => {
+            /* places centered (horizontally & vertically) text at x,y */
+            const metrics = ctx.measureText(ch);
+            const xx = x - metrics.width / 2;
+            const yy = y;
+            ctx.fillText(ch, xx, yy);
+        };
+        const v_place = (ch: string, x: number, y: number): void => {
+            /* places centered (horizontally & vertically) text at x,y, with text going down vertically. */
+            for (let i = 0; i < ch.length; ++i) {
+                const metrics = ctx.measureText(ch[i]);
+                const xx = x - metrics.width / 2;
+                let yy = y;
+                const H =
+                    metrics.width; /* should be height in an ideal world, measureText doesn't seem to return it though. For our purposes this works well enough though. */
 
-        if (force_clear || !this.grid_layer || !this.shadow_layer) {
-            this.shadow_layer?.remove();
-            this.shadow_layer = document.createElementNS("http://www.w3.org/2000/svg", "g");
-            this.shadow_layer.setAttribute("class", "shadow-layer");
-            this.svg.appendChild(this.shadow_layer);
-
-            this.grid_layer?.remove();
-            this.grid_layer = document.createElementNS("http://www.w3.org/2000/svg", "g");
-            this.grid_layer.setAttribute("class", "grid");
-            this.svg.appendChild(this.grid_layer);
-
-            for (let j = 0; j < this.height; ++j) {
-                this.grid[j] = [];
-                this.shadow_grid[j] = [];
-                /*
-                for (let i = 0; i < this.width; ++i) {
-                    const cell = document.createElementNS("http://www.w3.org/2000/svg", "g");
-                    cell.setAttribute("x", (i * this.square_size).toString());
-                    cell.setAttribute("y", (j * this.square_size).toString());
-                    cell.setAttribute("width", this.square_size.toString());
-                    cell.setAttribute("height", this.square_size.toString());
-                    this.grid_layer.appendChild(cell);
-
-                    this.grid[j][i] = cell;
+                if (ch.length === 2) {
+                    yy = yy - H + i * H;
                 }
-                */
+                if (ch.length === 3) {
+                    yy = yy - H * 1.5 + i * H;
+                }
+
+                ctx.fillText(ch[i], xx, yy);
             }
+        };
+
+        const drawHorizontal = (i: number, j: number): void => {
+            switch (this.getCoordinateDisplaySystem()) {
+                case "A1":
+                    for (let c = 0; c < this.width; ++i, ++c) {
+                        const x =
+                            (i -
+                                this.bounds.left -
+                                (this.bounds.left > 0 ? +this.draw_left_labels : 0)) *
+                                this.square_size +
+                            this.square_size / 2;
+                        const y = j * this.square_size + this.square_size / 2;
+                        place(GoMath.pretty_coor_num2ch(c), x, y);
+                    }
+                    break;
+                case "1-1":
+                    for (let c = 0; c < this.width; ++i, ++c) {
+                        const x =
+                            (i -
+                                this.bounds.left -
+                                (this.bounds.left > 0 ? +this.draw_left_labels : 0)) *
+                                this.square_size +
+                            this.square_size / 2;
+                        const y = j * this.square_size + this.square_size / 2;
+                        place("" + (c + 1), x, y);
+                    }
+                    break;
+            }
+        };
+
+        const drawVertical = (i: number, j: number): void => {
+            switch (this.getCoordinateDisplaySystem()) {
+                case "A1":
+                    for (let c = 0; c < this.height; ++j, ++c) {
+                        const x = i * this.square_size + this.square_size / 2;
+                        const y =
+                            (j -
+                                this.bounds.top -
+                                (this.bounds.top > 0 ? +this.draw_top_labels : 0)) *
+                                this.square_size +
+                            this.square_size / 2;
+                        place("" + (this.height - c), x, y);
+                    }
+                    break;
+                case "1-1":
+                    const chinese_japanese_numbers = [
+                        "一",
+                        "二",
+                        "三",
+                        "四",
+                        "五",
+                        "六",
+                        "七",
+                        "八",
+                        "九",
+                        "十",
+                        "十一",
+                        "十二",
+                        "十三",
+                        "十四",
+                        "十五",
+                        "十六",
+                        "十七",
+                        "十八",
+                        "十九",
+                        "二十",
+                        "二十一",
+                        "二十二",
+                        "二十三",
+                        "二十四",
+                        "二十五",
+                    ];
+                    for (let c = 0; c < this.height; ++j, ++c) {
+                        const x = i * this.square_size + this.square_size / 2;
+                        const y =
+                            (j -
+                                this.bounds.top -
+                                (this.bounds.top > 0 ? +this.draw_top_labels : 0)) *
+                                this.square_size +
+                            this.square_size / 2;
+                        v_place(chinese_japanese_numbers[c], x, y);
+                    }
+                    break;
+            }
+        };
+
+        if (force_clear || !this.__borders_initialized) {
+            this.__borders_initialized = true;
+            if (this.shadow_ctx) {
+                this.shadow_ctx.clearRect(0, 0, metrics.width, metrics.height);
+            }
+            ctx.clearRect(0, 0, metrics.width, metrics.height);
+
+            /* Draw labels */
+            let text_size = Math.round(this.square_size * 0.5);
+            let bold = "bold";
+            if (this.getCoordinateDisplaySystem() === "1-1") {
+                text_size *= 0.7;
+                bold = "";
+
+                if (this.height > 20) {
+                    text_size *= 0.7;
+                }
+            }
+
+            ctx.font = `${bold} ${text_size}px ${GOBAN_FONT}`;
+            ctx.textBaseline = "middle";
+            ctx.fillStyle = this.theme_board.getLabelTextColor();
+            ctx.save();
+
+            if (this.draw_top_labels && this.bounds.top === 0) {
+                drawHorizontal(this.draw_left_labels ? 1 : 0, 0);
+            }
+            if (this.draw_bottom_labels && this.bounds.bottom === this.height - 1) {
+                drawHorizontal(
+                    this.draw_left_labels ? 1 : 0,
+                    +this.draw_top_labels + this.bounded_height,
+                );
+            }
+            if (this.draw_left_labels && this.bounds.left === 0) {
+                drawVertical(0, this.draw_top_labels ? 1 : 0);
+            }
+            if (this.draw_right_labels && this.bounds.right === this.width - 1) {
+                drawVertical(+this.draw_left_labels + this.bounded_width, +this.draw_top_labels);
+            }
+
+            ctx.restore();
         }
 
         /* Draw squares */
@@ -2983,21 +2891,21 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             this.__draw_state = GoMath.makeStringMatrix(this.width, this.height);
         }
 
+        /* Set font for text overlay */
+        {
+            const text_size = Math.round(this.square_size * 0.45);
+            ctx.font = "bold " + text_size + "px " + GOBAN_FONT;
+        }
+
         for (let j = this.bounds.top; j <= this.bounds.bottom; ++j) {
             for (let i = this.bounds.left; i <= this.bounds.right; ++i) {
                 this.drawSquare(i, j);
             }
         }
 
-        if (this.pen_marks) {
-            if (force_clear) {
-                this.detachPenLayer();
-            }
-            this.drawPenMarks(this.pen_marks);
-        }
+        this.drawPenMarks(this.pen_marks);
         this.move_tree_redraw();
     }
-
     public showMessage(
         message_id_or_error: MessageID,
         parameters?: { [key: string]: any },
@@ -3056,7 +2964,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
     }
     public clearMessage(): void {
         if (this.message_div) {
-            this.message_div.remove();
+            this.message_div.parentNode?.removeChild(this.message_div);
             delete this.message_div;
         }
         if (this.message_timeout) {
@@ -3068,7 +2976,6 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
     }
     protected setThemes(themes: GobanSelectedThemes, dont_redraw: boolean): void {
         if (this.no_display) {
-            console.log("No display");
             return;
         }
 
@@ -3092,38 +2999,82 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             throw new Error("invalid stone radius computed");
         }
 
-        /*
+        if (
+            this.theme_white.stoneCastsShadow(this.theme_stone_radius) ||
+            this.theme_black.stoneCastsShadow(this.theme_stone_radius)
+        ) {
+            if (this.shadow_layer) {
+                resizeDeviceScaledCanvas(
+                    this.shadow_layer,
+                    this.metrics.width,
+                    this.metrics.height,
+                );
+                const ctx = this.shadow_layer.getContext("2d", { willReadFrequently: true });
+                if (ctx) {
+                    this.shadow_ctx = ctx;
+                } else {
+                    throw new Error(`Failed to get drawing context for shadow layer`);
+                }
+            } else {
+                this.attachShadowLayer();
+            }
+        } else {
+            this.detachShadowLayer();
+        }
+
+        if (!(themes.white in __theme_cache.white)) {
+            __theme_cache.white[themes.white] = {
+                creation_order: [],
+            };
+        }
+        if (!(themes.black in __theme_cache.black)) {
+            __theme_cache.black[themes.black] = {
+                creation_order: [],
+            };
+        }
+
         const deferredRenderCallback = () => {
             this.redraw(true);
             this.move_tree_redraw();
         };
-        */
 
         try {
-            this.svg_defs?.remove();
-            this.svg_defs = document.createElementNS("http://www.w3.org/2000/svg", "defs");
-            this.svg.appendChild(this.svg_defs);
-
-            if (!(themes.white in __theme_cache.white)) {
-                __theme_cache.white[themes.white] = {};
+            if (!(this.theme_stone_radius in __theme_cache.white[themes.white])) {
+                __theme_cache.white[themes.white][this.theme_stone_radius] =
+                    this.theme_white.preRenderWhite(
+                        this.theme_stone_radius,
+                        23434,
+                        deferredRenderCallback,
+                    );
+                __theme_cache.white[themes.white].creation_order.push(this.theme_stone_radius);
             }
-            if (!(themes.black in __theme_cache.black)) {
-                __theme_cache.black[themes.black] = {};
+            if (!(this.theme_stone_radius in __theme_cache.black[themes.black])) {
+                __theme_cache.black[themes.black][this.theme_stone_radius] =
+                    this.theme_black.preRenderBlack(
+                        this.theme_stone_radius,
+                        2081,
+                        deferredRenderCallback,
+                    );
+                __theme_cache.black[themes.black].creation_order.push(this.theme_stone_radius);
             }
 
-            for (const radius of [this.theme_stone_radius, MoveTree.stone_radius]) {
-                __theme_cache.white[themes.white][radius] = this.theme_white.preRenderWhiteSVG(
-                    this.svg_defs,
-                    radius,
-                    23434,
-                    () => {},
-                );
-                __theme_cache.black[themes.black][radius] = this.theme_black.preRenderBlackSVG(
-                    this.svg_defs,
-                    radius,
-                    2081,
-                    () => {},
-                );
+            if (!(MoveTree.stone_radius in __theme_cache.white[themes.white])) {
+                __theme_cache.white[themes.white][MoveTree.stone_radius] =
+                    this.theme_white.preRenderWhite(
+                        MoveTree.stone_radius,
+                        23434,
+                        deferredRenderCallback,
+                    );
+                __theme_cache.white[themes.white].creation_order.push(MoveTree.stone_radius);
+            }
+            if (!(MoveTree.stone_radius in __theme_cache.black[themes.black])) {
+                __theme_cache.black[themes.black][MoveTree.stone_radius] =
+                    this.theme_black.preRenderBlack(
+                        MoveTree.stone_radius,
+                        2081,
+                        deferredRenderCallback,
+                    );
+                __theme_cache.black[themes.black].creation_order.push(MoveTree.stone_radius);
             }
         } catch (e) {
             console.error(`Error pre-rendering stones.`, {
@@ -3131,6 +3082,41 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 move_tree_stone_radius: MoveTree.stone_radius,
             });
             throw e;
+        }
+
+        // We should only need a few sizes, like 6 in most cases, but when we resize a window slowly or
+        // have a bunch of weird sized boards, we'll need more. These are very small and there aren't
+        // any devices that should have a problem with them, except for an artificial limit on iOS devices
+        // which we'll handle below.
+        let max_cache_size = 500;
+        try {
+            /* ipads only allow a very small amount of memory to be allocated to canvases,
+             * so we will be more aggressive about cleaning up the cache on those devices */
+            if (
+                /iP(ad|hone|od).+(Version\/[\d.]|OS \d.*like mac os x)+.*Safari/i.test(
+                    navigator.userAgent,
+                )
+            ) {
+                console.log("iOS device detected, reducing cache size");
+                max_cache_size = 12; // mini goban, main boards, 9x9, 13x13, 19x19 should account for 6. We double that for good measure for odd sizes and resizing.
+            }
+        } catch (e) {
+            console.error(e);
+        }
+
+        if (__theme_cache.black[themes.black].creation_order.length > max_cache_size) {
+            const old_radius = __theme_cache.black[themes.black].creation_order.shift();
+            if (old_radius) {
+                console.log("deleting old radius [black]", old_radius);
+                delete __theme_cache.black[themes.black][old_radius];
+            }
+        }
+        if (__theme_cache.white[themes.white].creation_order.length > max_cache_size) {
+            const old_radius = __theme_cache.white[themes.white].creation_order.shift();
+            if (old_radius) {
+                console.log("deleting old radius [white]", old_radius);
+                delete __theme_cache.white[themes.white][old_radius];
+            }
         }
 
         this.theme_white_stones = __theme_cache.white[themes.white][this.theme_stone_radius];
@@ -3147,6 +3133,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         this.theme_blank_text_color = this.theme_board.getBlankTextColor();
         this.theme_black_text_color = this.theme_black.getBlackTextColor();
         this.theme_white_text_color = this.theme_white.getWhiteTextColor();
+        //this.parent.css(this.theme_board.getBackgroundCSS());
         const bg_css = this.theme_board.getBackgroundCSS();
         if (this.parent) {
             for (const key in bg_css) {
@@ -3160,7 +3147,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         }
     }
     private onLabelingStart(ev: MouseEvent | TouchEvent) {
-        const pos = getRelativeEventPosition(ev, this.parent);
+        const pos = getRelativeEventPosition(ev);
         this.last_label_position = this.xy2ij(pos.x, pos.y, false);
 
         {
@@ -3190,7 +3177,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         this.drawSquare(this.last_label_position.i, this.last_label_position.j);
     }
     private onLabelingMove(ev: MouseEvent | TouchEvent) {
-        const pos = getRelativeEventPosition(ev, this.parent);
+        const pos = getRelativeEventPosition(ev);
         const cur = this.xy2ij(pos.x, pos.y);
 
         {
@@ -3207,6 +3194,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             this.setLabelCharacterFromMarks();
         }
     }
+
     private onAnalysisScoringStart(ev: MouseEvent | TouchEvent) {
         const pos = getRelativeEventPosition(ev, this.parent);
         this.analysis_scoring_last_position = this.xy2ij(pos.x, pos.y, false);
@@ -3302,13 +3290,12 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         if (!this.move_tree_inner_container) {
             do_init = true;
             this.move_tree_inner_container = document.createElement("div");
-            //this.move_tree_canvas = allocateCanvasOrError();
-            this.move_tree_svg = document.createElementNS("http://www.w3.org/2000/svg", "svg");
-            this.move_tree_inner_container.appendChild(this.move_tree_svg);
+            this.move_tree_canvas = allocateCanvasOrError();
+            this.move_tree_inner_container.appendChild(this.move_tree_canvas);
             this.move_tree_container.appendChild(this.move_tree_inner_container);
-            this.move_tree_bindEvents(this.move_tree_svg);
+            this.move_tree_bindCanvasEvents(this.move_tree_canvas);
             this.move_tree_container.style.position = "relative";
-            this.move_tree_svg.style.position = "absolute";
+            this.move_tree_canvas.style.position = "absolute";
 
             try {
                 const observer = new ResizeObserver(() => {
@@ -3323,8 +3310,8 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             }
         }
 
-        if (!this.move_tree_svg) {
-            console.warn(`move_tree_redraw called without move_tree_svg set`);
+        if (!this.move_tree_canvas) {
+            console.warn(`move_tree_redraw called without move_tree_canvas set`);
             return;
         }
 
@@ -3347,19 +3334,18 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             });
         }
 
-        if (
-            this.move_tree_svg.clientWidth !== this.move_tree_container.clientWidth ||
-            this.move_tree_svg.clientHeight !== this.move_tree_container.clientHeight
+        /*
+        if (this.move_tree_canvas.width !== this.move_tree_container.outerWidth ||
+            this.move_tree_canvas.height !== this.move_tree_container.outerHeight
         ) {
-            this.move_tree_svg.setAttribute(
-                "width",
-                this.move_tree_container.clientWidth.toString(),
-            );
-            this.move_tree_svg.setAttribute(
-                "height",
-                this.move_tree_container.clientHeight.toString(),
-            );
+            console.log(this.move_tree_canvas.width, this.move_tree_container.outerWidth,
+                this.move_tree_canvas.height, this.move_tree_container.outerHeight);
+            this.move_tree_canvas.width = this.move_tree_container.outerWidth;
+            this.move_tree_canvas.height = this.move_tree_container.outerHeight;
+            this.move_tree_canvas.style.width = this.move_tree_container.outerWidth + "px";
+            this.move_tree_canvas.style.height = this.move_tree_container.outerHeight + "px";
         }
+        */
 
         this.engine.move_tree.recomputeIsobranches();
         const active_path_end = this.engine.cur_move;
@@ -3368,8 +3354,18 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
         active_path_end.setActivePath(++MoveTree.active_path_number);
 
-        //const canvas = this.move_tree_canvas;
-        const svg = this.move_tree_svg;
+        /*
+        if (!this.move_tree_container.data("move-tree-redraw-on-scroll")) {
+            let debounce = false;
+            this.redraw_on_scroll = () => {
+                MoveTree.redraw_root.redraw(MoveTree.redraw_config, true);
+            };
+            this.move_tree_container.data("move-tree-redraw-on-scroll", this.redraw_on_scroll);
+            this.move_tree_container.scroll(this.redraw_on_scroll);
+        }
+        */
+
+        const canvas = this.move_tree_canvas;
         const engine = this.engine;
 
         this.engine.move_tree_layout_vector = [];
@@ -3392,6 +3388,14 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         let div_scroll_top = this.move_tree_container.scrollTop;
         let div_scroll_left = this.move_tree_container.scrollLeft;
 
+        if (canvas.width !== div_clientWidth || canvas.height !== div_clientHeight) {
+            resizeDeviceScaledCanvas(
+                canvas,
+                this.move_tree_container.clientWidth,
+                this.move_tree_container.clientHeight,
+            );
+        }
+
         this.move_tree_inner_container.style.width = width + "px";
         this.move_tree_inner_container.style.height = height + "px";
 
@@ -3412,8 +3416,8 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             }
         }
 
-        svg.style.top = div_scroll_top + "px";
-        svg.style.left = div_scroll_left + "px";
+        canvas.style.top = div_scroll_top + "px";
+        canvas.style.left = div_scroll_left + "px";
 
         const viewport = {
             offset_x: div_scroll_left,
@@ -3424,24 +3428,38 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             maxy: div_scroll_top + div_clientHeight + MoveTree.stone_square_size,
         };
 
-        svg.innerHTML = "";
+        const ctx = canvas.getContext("2d", { willReadFrequently: true });
+        if (!ctx) {
+            throw new Error(`Failed to get drawing context for move tree canvas`);
+        }
+        ctx.clearRect(0, 0, canvas.width, canvas.height);
 
-        this.move_tree_hilightNode(svg, active_path_end, "#6BAADA", viewport);
+        this.move_tree_hilightNode(ctx, active_path_end, "#6BAADA", viewport);
 
         if (engine.cur_review_move && engine.cur_review_move.id !== active_path_end.id) {
-            this.move_tree_hilightNode(svg, engine.cur_review_move, "#6BDA6B", viewport);
+            this.move_tree_hilightNode(ctx, engine.cur_review_move, "#6BDA6B", viewport);
         }
 
-        this.move_tree_recursiveDrawPath(svg, this.engine.move_tree, viewport);
+        ctx.save();
+        ctx.lineWidth = 1.0;
+        ctx.strokeStyle = this.theme_line_color;
+        this.move_tree_recursiveDrawPath(ctx, this.engine.move_tree, viewport);
+        ctx.restore();
 
+        ctx.save();
+        ctx.globalCompositeOperation = "source-over";
+        const text_size = 10;
+        ctx.font = `bold ${text_size}px Verdana,Arial,sans-serif`;
+        ctx.textBaseline = "middle";
         this.move_tree_drawRecursive(
-            svg,
+            ctx,
             this.engine.move_tree,
             MoveTree.active_path_number,
             viewport,
         );
+        ctx.restore();
     }
-    public move_tree_bindEvents(svg: SVGElement): void {
+    public move_tree_bindCanvasEvents(canvas: HTMLCanvasElement): void {
         const handler = (event: TouchEvent | MouseEvent) => {
             try {
                 if (!this.move_tree_container) {
@@ -3450,7 +3468,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
                 const ox = this.move_tree_container.scrollLeft;
                 const oy = this.move_tree_container.scrollTop;
-                const pos = getRelativeEventPosition(event, this.move_tree_container);
+                const pos = getRelativeEventPosition(event);
                 pos.x += ox;
                 pos.y += oy;
                 const i = Math.floor(pos.x / MoveTree.stone_square_size);
@@ -3471,7 +3489,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                         // nasty looking things to do with Touch etc, so using it here
                         // gets around that kind of thing, even though in theory it
                         // might be nicer to sent the client absolute coords, maybe.
-                        const rpos = getRelativeEventPosition(event, this.move_tree_container);
+                        const rpos = getRelativeEventPosition(event);
                         this.emit("played-by-click", {
                             player_id: this.engine.cur_move.played_by,
                             x: rpos.x,
@@ -3484,17 +3502,29 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             }
         };
 
-        svg.addEventListener("touchstart", handler);
-        svg.addEventListener("mousedown", handler);
+        canvas.addEventListener("touchstart", handler);
+        canvas.addEventListener("mousedown", handler);
 
         this.on("destroy", () => {
-            svg.removeEventListener("touchstart", handler);
-            svg.removeEventListener("mousedown", handler);
+            canvas.removeEventListener("touchstart", handler);
+            canvas.removeEventListener("mousedown", handler);
         });
     }
 
+    protected computeThemeStoneRadius(): number {
+        // Scale proportionally in general
+        let r = this.square_size * 0.488;
+
+        // Prevent pixel sharing in low-res
+        if (this.square_size % 2 === 0) {
+            r = Math.min(r, (this.square_size - 1) / 2);
+        }
+
+        return Math.max(1, r);
+    }
+
     move_tree_drawStone(
-        svg: SVGElement,
+        ctx: CanvasRenderingContext2D,
         node: MoveTree,
         active_path_number: number,
         viewport: ViewPortInterface,
@@ -3504,19 +3534,14 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
         const cy = node.layout_cy - viewport.offset_y;
         const color = node.player;
         const on_path = node.active_path_number === active_path_number;
-        const r = MoveTree.stone_radius;
 
-        const cell = document.createElementNS("http://www.w3.org/2000/svg", "g");
-        svg.appendChild(cell);
         if (!on_path) {
-            cell.setAttribute("class", "move-tree-stone");
-            cell.setAttribute("fill-opacity", "0.6");
-            cell.setAttribute("stroke-opacity", "0.6");
-            cell.setAttribute("opacity", "0.6");
+            ctx.save();
+            ctx.globalAlpha = 0.4;
         }
 
-        const theme_white_stones = __theme_cache.white[this.themes.white][r];
-        const theme_black_stones = __theme_cache.black[this.themes.black][r];
+        const theme_white_stones = __theme_cache.white[this.themes.white][MoveTree.stone_radius];
+        const theme_black_stones = __theme_cache.black[this.themes.black][MoveTree.stone_radius];
 
         if (!theme_white_stones || !theme_black_stones) {
             throw new Error(
@@ -3526,10 +3551,10 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
         if (color === 1) {
             const stone = theme_black_stones[stone_idx % theme_black_stones.length];
-            this.theme_black.placeBlackStoneSVG(cell, undefined, stone, cx, cy, r);
+            this.theme_black.placeBlackStone(ctx, null, stone, cx, cy, MoveTree.stone_radius);
         } else if (color === 2) {
             const stone = theme_white_stones[stone_idx % theme_white_stones.length];
-            this.theme_white.placeWhiteStoneSVG(cell, undefined, stone, cx, cy, r);
+            this.theme_white.placeWhiteStone(ctx, null, stone, cx, cy, MoveTree.stone_radius);
         } else {
             return;
         }
@@ -3561,52 +3586,55 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             delete node.label_metrics;
         }
 
-        const font_size = 10;
-        const text = document.createElementNS("http://www.w3.org/2000/svg", "text");
-        text.setAttribute("x", cx.toFixed(0));
-        text.setAttribute("y", (cy + font_size / 8).toFixed(1));
-        text.setAttribute("width", MoveTree.stone_square_size.toFixed(1));
-        text.setAttribute("height", MoveTree.stone_square_size.toFixed(1));
-        text.setAttribute("font-size", `${font_size}px`);
-        text.setAttribute("font-weight", "bold");
-        text.setAttribute("alignment-baseline", "middle");
-        text.setAttribute("text-anchor", "middle");
-        text.setAttribute("dominant-baseline", "middle");
-        text.setAttribute("fill", text_color);
-        text.textContent = node.label;
+        ctx.fillStyle = text_color;
+        //ctx.strokeStyle=text_outline_color;
+        if (!node.label_metrics) {
+            node.label_metrics = ctx.measureText(node.label);
+        }
+        const metrics = node.label_metrics;
+        const xx = cx - metrics.width / 2;
+        const yy =
+            cy +
+            (/WebKit|Trident/.test(navigator.userAgent)
+                ? MoveTree.stone_radius * -0.01
+                : 1); /* middle centering is different on firefox */
+        //ctx.strokeText(node.label, xx, yy);
+        ctx.fillText(node.label, xx, yy);
 
-        cell.appendChild(text);
+        if (!on_path) {
+            ctx.restore();
+        }
 
-        const ring_color = node.text
-            ? "#3333ff"
-            : node.correct_answer
-              ? "#33ff33"
-              : node.wrong_answer
-                ? "#ff3333"
-                : null;
+        let ring_color = null;
 
+        if (node.text) {
+            ring_color = "#3333ff";
+        }
+        if (node.correct_answer) {
+            ring_color = "#33ff33";
+        }
+        if (node.wrong_answer) {
+            ring_color = "#ff3333";
+        }
         if (ring_color) {
-            const ring = document.createElementNS("http://www.w3.org/2000/svg", "circle");
-            ring.setAttribute("cx", cx.toFixed(0));
-            ring.setAttribute("cy", cy.toFixed(1));
-            ring.setAttribute("r", r.toFixed(0));
-            ring.setAttribute("fill", "none");
-            ring.setAttribute("stroke", ring_color);
-            ring.setAttribute("stroke-width", "2");
-            cell.appendChild(ring);
+            ctx.beginPath();
+            ctx.strokeStyle = ring_color;
+            ctx.lineWidth = 2.0;
+            ctx.arc(cx, cy, MoveTree.stone_radius, 0, 2 * Math.PI, true);
+            ctx.stroke();
         }
     }
     move_tree_drawRecursive(
-        svg: SVGElement,
+        ctx: CanvasRenderingContext2D,
         node: MoveTree,
         active_path_number: number,
         viewport: ViewPortInterface,
     ): void {
         if (node.trunk_next) {
-            this.move_tree_drawRecursive(svg, node.trunk_next, active_path_number, viewport);
+            this.move_tree_drawRecursive(ctx, node.trunk_next, active_path_number, viewport);
         }
         for (let i = 0; i < node.branches.length; ++i) {
-            this.move_tree_drawRecursive(svg, node.branches[i], active_path_number, viewport);
+            this.move_tree_drawRecursive(ctx, node.branches[i], active_path_number, viewport);
         }
 
         if (
@@ -3616,29 +3644,30 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 node.layout_cy >= viewport.miny &&
                 node.layout_cy <= viewport.maxy)
         ) {
-            this.move_tree_drawStone(svg, node, active_path_number, viewport);
+            this.move_tree_drawStone(ctx, node, active_path_number, viewport);
         }
     }
     move_tree_hilightNode(
-        svg: SVGElement,
+        ctx: CanvasRenderingContext2D,
         node: MoveTree,
         color: string,
         viewport: ViewPortInterface,
     ): void {
-        const rect = document.createElementNS("http://www.w3.org/2000/svg", "rect");
+        ctx.beginPath();
         const sx =
             Math.round(node.layout_cx - MoveTree.stone_square_size * 0.5) - viewport.offset_x;
         const sy =
             Math.round(node.layout_cy - MoveTree.stone_square_size * 0.5) - viewport.offset_y;
-        rect.setAttribute("x", sx.toFixed(0));
-        rect.setAttribute("y", sy.toFixed(0));
-        rect.setAttribute("width", MoveTree.stone_square_size.toFixed(0));
-        rect.setAttribute("height", MoveTree.stone_square_size.toFixed(0));
-        rect.setAttribute("fill", color);
-        svg.appendChild(rect);
+        ctx.rect(sx, sy, MoveTree.stone_square_size, MoveTree.stone_square_size);
+        ctx.fillStyle = color;
+        ctx.fill();
     }
 
-    move_tree_drawPath(svg: SVGElement, node: MoveTree, viewport: ViewPortInterface): void {
+    move_tree_drawPath(
+        ctx: CanvasRenderingContext2D,
+        node: MoveTree,
+        viewport: ViewPortInterface,
+    ): void {
         if (node.parent) {
             if (node.parent.layout_cx < viewport.minx && node.layout_cx < viewport.minx) {
                 return;
@@ -3653,26 +3682,22 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
                 return;
             }
 
-            const curve = document.createElementNS("http://www.w3.org/2000/svg", "path");
+            ctx.beginPath();
+            ctx.strokeStyle = node.trunk ? "#000000" : MoveTree.line_colors[node.line_color];
             const ox = viewport.offset_x;
             const oy = viewport.offset_y;
-            curve.setAttribute(
-                "d",
-                `M ${node.parent.layout_cx - ox} ${node.parent.layout_cy - oy} Q ${
-                    node.layout_cx - MoveTree.stone_square_size * 0.5 - ox
-                } ${node.layout_cy - oy} ${node.layout_cx - ox} ${node.layout_cy - oy}`,
+            ctx.moveTo(node.parent.layout_cx - ox, node.parent.layout_cy - oy);
+            ctx.quadraticCurveTo(
+                node.layout_cx - MoveTree.stone_square_size * 0.5 - ox,
+                node.layout_cy - oy,
+                node.layout_cx - ox,
+                node.layout_cy - oy,
             );
-            curve.setAttribute("fill", "none");
-            curve.setAttribute(
-                "stroke",
-                node.trunk ? "#000000" : MoveTree.line_colors[node.line_color],
-            );
-            curve.setAttribute("stroke-width", "1");
-            svg.appendChild(curve);
+            ctx.stroke();
         }
     }
     move_tree_drawIsoBranchTo(
-        svg: SVGElement,
+        ctx: CanvasRenderingContext2D,
         from_node: MoveTree,
         to_node: MoveTree,
         viewport: ViewPortInterface,
@@ -3694,6 +3719,13 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             return;
         }
 
+        /*
+        let isStrong = (a, b):boolean => {
+            return a.trunk_next === null && a.branches.length === 0 && (b.trunk_next != null || b.branches.length !== 0);
+        };
+        */
+
+        // isStrong(B, A)) {
         if (
             B.trunk_next === null &&
             B.branches.length === 0 &&
@@ -3704,6 +3736,7 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
             B = t;
         }
 
+        //isStrong(A, B);
         const strong =
             A.trunk_next == null &&
             A.branches.length === 0 &&
@@ -3711,39 +3744,85 @@ export class GobanSVG extends GobanCore implements GobanSVGInterface {
 
         const ox = viewport.offset_x;
         const oy = viewport.offset_y;
-
-        const curve = document.createElementNS("http://www.w3.org/2000/svg", "path");
+        ctx.beginPath();
+        ctx.strokeStyle = MoveTree.isobranch_colors[strong ? "strong" : "weak"];
+        const cur_line_width = ctx.lineWidth;
+        ctx.lineWidth = 2;
+        ctx.moveTo(B.layout_cx - ox, B.layout_cy - oy);
         const my = strong ? B.layout_cy : (A.layout_cy + B.layout_cy) / 2;
         const mx = (A.layout_cx + B.layout_cx) / 2 + MoveTree.stone_square_size * 0.5;
-        curve.setAttribute(
-            "d",
-            `M ${B.layout_cx - ox} ${B.layout_cy - oy} Q ${mx - ox} ${my - oy} ${
-                A.layout_cx - ox
-            } ${A.layout_cy - oy}`,
-        );
-        curve.setAttribute("fill", "none");
-        curve.setAttribute("stroke", MoveTree.isobranch_colors[strong ? "strong" : "weak"]);
-        curve.setAttribute("stroke-width", "2");
-        svg.appendChild(curve);
+        ctx.quadraticCurveTo(mx - ox, my - oy, A.layout_cx - ox, A.layout_cy - oy);
+        ctx.stroke();
+        ctx.lineWidth = cur_line_width;
     }
     move_tree_recursiveDrawPath(
-        svg: SVGElement,
+        ctx: CanvasRenderingContext2D,
         node: MoveTree,
         viewport: ViewPortInterface,
     ): void {
         if (node.trunk_next) {
-            this.move_tree_recursiveDrawPath(svg, node.trunk_next, viewport);
+            this.move_tree_recursiveDrawPath(ctx, node.trunk_next, viewport);
         }
         for (let i = 0; i < node.branches.length; ++i) {
-            this.move_tree_recursiveDrawPath(svg, node.branches[i], viewport);
+            this.move_tree_recursiveDrawPath(ctx, node.branches[i], viewport);
         }
 
         if (node.isobranches) {
             for (let i = 0; i < node.isobranches.length; ++i) {
-                this.move_tree_drawIsoBranchTo(svg, node, node.isobranches[i], viewport);
+                this.move_tree_drawIsoBranchTo(ctx, node, node.isobranches[i], viewport);
             }
         }
 
-        this.move_tree_drawPath(svg, node, viewport);
+        /* only consider x, since lines can extend awhile on the y */
+        //if (this.layout_cx >= viewport.minx && this.layout_cx <= viewport.maxx) {
+        this.move_tree_drawPath(ctx, node, viewport);
+        //}
     }
+}
+
+const fitTextCache: { [key: string]: [number, string, TextMetrics] } = {};
+
+// fontPattern MUST have a FONT_SIZE string in it that will be replaced
+// ctx.font will be set to the appropriate size when this returns.
+// @returns font_size, font_string, metrics
+function fitText(
+    ctx: CanvasRenderingContext2D,
+    text: string,
+    fontPattern: string,
+    startingFontSize: number,
+    width: number,
+): [number, string, TextMetrics] {
+    const MIN_FONT_SIZE = 4;
+
+    if (fontPattern.indexOf("FONT_SIZE") < 0) {
+        throw new Error(
+            `fitText expects FONT_SIZE to be present in the fontPattern, was ${fontPattern}`,
+        );
+    }
+
+    const key = `${width} ${fontPattern} ${startingFontSize} ${text}`;
+    if (key in fitTextCache) {
+        const cached = fitTextCache[key];
+        ctx.font = cached[1];
+        return cached;
+    }
+
+    let font_size = startingFontSize;
+
+    do {
+        const font = fontPattern.replace("FONT_SIZE", font_size.toString());
+        ctx.font = font;
+        const metrics = ctx.measureText(text);
+        if (font_size <= MIN_FONT_SIZE || metrics.width < width) {
+            fitTextCache[key] = [font_size, font, metrics];
+            return fitTextCache[key];
+        }
+
+        const new_font_size = Math.floor((font_size * width) / metrics.width);
+        if (new_font_size >= font_size) {
+            font_size = font_size - 1;
+        } else {
+            font_size = new_font_size;
+        }
+    } while (true);
 }
