+++ tsconfig.node.json
@@ -13,7 +13,8 @@
         "lib": ["es2017", "dom"],
         "baseUrl": ".",
         "paths": {
-            "*": ["src/*", "*"]
+            "*": ["src/*", "*"],
+            "goscorer": ["src/third_party/goscorer/goscorer"]
         },
         "noImplicitAny": true,
         "noImplicitReturns": true,
